% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Add.R
\docType{methods}
\name{addExecOutput,DArch-method}
\alias{addExecOutput,DArch-method}
\title{Adds an execution output for a DArch object}
\usage{
\S4method{addExecOutput}{DArch}(darch, output)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}

\item{output}{The output of the layer.}
}
\description{
Adds an execution output for a DArch object
}
\seealso{
\link{addExecOutput}
}
\keyword{internal}

