% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Add.R
\name{addExecOutput}
\alias{addExecOutput}
\title{Adds an execution output for a DArch object}
\usage{
addExecOutput(darch, output)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}

\item{output}{The output of the layer.}
}
\description{
This method can be used to save the execution outputs of every layer for the
DArch object. The outputs are saved in a list and every time this function is
called, the list is extended of one field with the new output.
}
\seealso{
\code{\linkS4class{DArch}}
}
\keyword{internal}

