% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Add.R
\docType{methods}
\name{addLayer,DArch-method}
\alias{addLayer,DArch-method}
\title{Adds a layer to the \code{\linkS4class{DArch}} object}
\usage{
\S4method{addLayer}{DArch}(darch, weights, biases, unitFunction)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}

\item{weights}{The weights for the layer.}

\item{biases}{The biases for the layer.}

\item{unitFunction}{The functions of the units in the layer.}
}
\description{
Adds a layer to the \code{\linkS4class{DArch}} object
}
\seealso{
\code{\link{addLayer}}
}
\keyword{internal}

