% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darch.Add.R
\name{addLayer}
\alias{addLayer}
\title{Adds a layer to the \code{\linkS4class{DArch}} object}
\usage{
addLayer(darch, weights, biases, unitFunction, weightUpdateFunction)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}

\item{weights}{The weights for the layer.}

\item{biases}{The biases for the layer.}

\item{unitFunction}{The functions of the units in the layer.}
}
\description{
Adds a layer to the given \code{\linkS4class{DArch}} object. The parameter
weights and biases will be put together in one matrix.
}
\seealso{
\code{\linkS4class{DArch}}
}
\keyword{internal}

