% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Add.R
\docType{methods}
\name{addLayerField,DArch-method}
\alias{addLayerField,DArch-method}
\title{Adds a field to a layer}
\usage{
\S4method{addLayerField}{DArch}(darch, index, field)
}
\arguments{
\item{darch}{An instance of the class \code{\link{DArch}}.}

\item{index}{The position of the layer.}

\item{field}{The new field for the layer.}
}
\description{
Adds a field to a layer
}
\seealso{
\code{\link{addLayerField}}
}
\keyword{internal}

