\docType{methods}
\name{addLayerField}
\alias{addLayerField}
\alias{addLayerField,DArch-method}
\title{Adds a field to a layer}
\usage{
addLayerField(darch, index, field)

\S4method{addLayerField}{DArch}(darch, index, field)
}
\arguments{
  \item{darch}{An instance of the class
  \code{\link{DArch}}.}

  \item{index}{The position of the layer.}

  \item{field}{The new field for the layer.}
}
\description{
Adds a field to the layer given by the index.
}
\seealso{
\code{\link{DArch}}
}

