% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Add.R
\name{addLayerField}
\alias{addLayerField}
\title{Adds a field to a layer}
\usage{
addLayerField(darch, index, field)
}
\arguments{
\item{darch}{An instance of the class \code{\link{DArch}}.}

\item{index}{The position of the layer.}

\item{field}{The new field for the layer.}
}
\description{
Adds a field to the layer given by the index.
}
\seealso{
\code{\linkS4class{DArch}}
}
\keyword{internal}

