% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dropout.R
\name{applyDropoutMask}
\alias{applyDropoutMask}
\title{Applies the given dropout mask to the given data row-wise.}
\usage{
applyDropoutMask(data, mask)
}
\arguments{
\item{data}{Data to which the dropout mask should be applied}

\item{mask}{The dropout mask, a vector of 0 and 1.}
}
\value{
Data with applied dropout mask
}
\description{
Applies the given dropout mask to the given data row-wise.
}

