% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{binSigmoidUnit}
\alias{binSigmoidUnit}
\title{Binary sigmoid unit function.}
\usage{
binSigmoidUnit(data, weights)
}
\arguments{
\item{data}{The data matrix for the calculation}

\item{weights}{The weight and bias matrix for the calculation}
}
\value{
A list with the binary activation of the unit in the first entry.
}
\description{
The function calculates the activation and the output from the sigmoid
transfer function. It returns a binary matrix where a entry is 1 if the value
is bigger than a random number generated with \code{\link{runif}}.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{linearUnitDerivative}};
  \code{\link{linearUnit}};
  \code{\link{maxoutUnitDerivative}};
  \code{\link{sigmoidUnitDerivative}};
  \code{\link{sigmoidUnit}};
  \code{\link{softmaxUnitDerivative}};
  \code{\link{softmaxUnit}};
  \code{\link{tanSigmoidUnitDerivative}};
  \code{\link{tanSigmoidUnit}}
}

