% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{createDataSet,ANY,missing,formula,missing-method}
\alias{createDataSet,ANY,missing,formula,missing-method}
\title{Create \code{\linkS4class{DataSet}} from \code{\link{formula}}.}
\usage{
\S4method{createDataSet}{ANY,missing,formula,missing}(data, formula, ...,
  subset, na.action, contrasts = NULL, scale = F)
}
\arguments{
\item{data}{Input data, possibly also target data if a formula is used.}

\item{formula}{Model \code{\link{formula}}.}
}
\description{
Create \code{\linkS4class{DataSet}} from \code{\link{formula}}.
}

