% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDataSet.default}
\alias{createDataSet.default}
\title{Create \code{\linkS4class{DataSet}} using data and targets.}
\usage{
createDataSet.default(data, targets, ...)
}
\arguments{
\item{data}{Input data, possibly also target data if a formula is used.}

\item{targets}{Target data.}
}
\description{
Create \code{\linkS4class{DataSet}} using data and targets.
}

