% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDataSet.formula}
\alias{createDataSet.formula}
\title{Constructor function for \code{\linkS4class{DataSet}} objects.}
\usage{
createDataSet.formula(data, formula, ..., subset, na.action, contrasts = NULL)
}
\arguments{
\item{data}{Data frame or matrix.}

\item{formula}{The formula specifying the model.}

\item{...}{additional parameters}

\item{subset}{Row indexing vector, \strong{not} parameter to
\code{\link{model.frame}}}

\item{na.action}{\code{\link{model.frame}} parameter}

\item{contrasts}{\code{\link{model.frame}} parameter}
}
\value{
The new \code{\linkS4class{DataSet}} object
}
\description{
Generates a new \code{\linkS4class{DataSet}} object with the given
parameters.
}
\details{
Initializes a \code{\linkS4class{DataSet}} with data using a
  \code{\link{formula}} and saving its parameters for subsequent data.
}
\seealso{
\code{\link{darch.formula}}
}

