% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDataSet}
\alias{createDataSet}
\title{Create data set using data, targets, a formula, and possibly an existing data
set.}
\usage{
createDataSet(data, targets, formula, dataSet, ...)
}
\arguments{
\item{data}{Input data, possibly also target data if a formula is used.}

\item{targets}{Target data.}

\item{formula}{Model \code{\link{formula}}.}

\item{dataSet}{\code{\linkS4class{DataSet}} to be used as the basis for the
new one}

\item{...}{Further parameters.}
}
\value{
New \code{\linkS4class{DataSet}}
}
\description{
Create data set using data, targets, a formula, and possibly an existing data
set.
}
\seealso{
\link{createDataSet.default}, \link{createDataSet.formula},
 \link{createDataSet.DataSet}
}

