% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/math.R
\name{gpuMatMult}
\alias{gpuMatMult}
\title{Helper function for matrix multiplication to keep gputools an optional
dependency.}
\usage{
gpuMatMult(m1, m2)
}
\arguments{
\item{m1}{First matrix}

\item{m2}{Second matrix}
}
\description{
Helper function for matrix multiplication to keep gputools an optional
dependency.
}

