% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{darch.DataSet}
\alias{darch.DataSet}
\title{Create and train DArch object using a \code{\linkS4class{DataSet}}.}
\usage{
\method{darch}{DataSet}(dataSet, ...)
}
\arguments{
\item{dataSet}{\code{\linkS4class{DataSet}}}
}
\value{
Fitted \code{\linkS4class{DArch}} instance
}
\description{
Convencience method which calls \code{\link{darch.default}}
}
\seealso{
\code{\link{darch.default}}
}

