% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{darch.default}
\alias{darch.default}
\title{Fit deep neural network with optional pre-training and fine-tuning.}
\usage{
\method{darch}{default}(x, y, darch.layers, rbm.learnRateWeights = 0.1,
  rbm.learnRateBiasVisible = 0.1, rbm.learnRateBiasHidden = 0.1,
  rbm.weightCost = 2e-04, rbm.momentum = 0.9, rbm.finalMomentum = 0.5,
  rbm.momentumSwitch = 5, rbm.visibleUnitFunction = sigmUnitFunc,
  rbm.hiddenUnitFunction = sigmUnitFuncSwitch,
  rbm.updateFunction = rbmUpdate, rbm.errorFunction = mseError,
  rbm.genWeightFunction = generateWeights, rbm.numCD = 1,
  rbm.maxEpoch = 0, darch = NULL, darch.batchSize = 1,
  darch.genWeightFunc = generateWeights, darch.logLevel = INFO,
  darch.fineTuneFunction = backpropagation, darch.momentum = 0.9,
  darch.finalMomentum = 0.5, darch.momentumSwitch = 5,
  darch.learnRateWeights = 0.001, darch.learnRateBiases = 0.001,
  darch.errorFunction = mseError, darch.dropoutInput = 0,
  darch.dropoutHidden = 0,
  darch.layerFunctionDefault = linearUnitDerivative,
  darch.layerFunctions = list(),
  darch.layerFunction.maxout.poolSize = getOption("darch.unitFunction.maxout.poolSize",
  NULL), darch.isBin = F, darch.isClass = T, darch.stopErr = -Inf,
  darch.stopClassErr = 101, darch.stopValidErr = -Inf,
  darch.stopValidClassErr = 101, darch.maxEpoch = 0, dataSetTrain = NULL,
  additionalDataSets = list())
}
\value{
Fitted \code{\linkS4class{DArch}} instance
}
\description{
TODO documentation parameters / details
}

