% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{darch.formula}
\alias{darch.formula}
\title{Fit a deep neural network using a formula and a single data frame or matrix.}
\usage{
\method{darch}{formula}(formula, data, ...)
}
\arguments{
\item{formula}{The formula specifying the model.}

\item{data}{Data frame or matrix.}

\item{...}{additional parameters}
}
\description{
Fit a deep neural network using a formula and a single data frame or matrix.
}
\seealso{
\code{\link{model.frame}}

\code{\link{createDataSet.formula}}
}

