\name{DArch}
\alias{DArch}
\alias{DArch-class}
\title{Class for deep architectures}
\description{
  To the constructor function: \code{\link{newDArch}} TODO:
  Description ...
}
\details{
  TODO: Details ...
}
\section{Slots}{
  \describe{ \item{\code{rbmList}:}{Object of class
  \code{"list"}. Contains all rbm's for the pre-training.}
  \item{\code{layers}:}{Object of class \code{"list"}. In
  the first field are the weights and in the second field
  is the unit function. }
  \item{\code{learnRateBiases}:}{Object of class
  \code{"list"}. Contains }
  \item{\code{fineTuneFunction}:}{Object of class
  \code{"function"}. }
  \item{\code{executeFunction}:}{Object of class
  \code{"list"}. } \item{\code{executeOutput}:}{Object of
  class \code{"list"}. } }
}
\author{
  Martin Drees
}
\seealso{
  \code{\linkS4class{Net}}, \code{\linkS4class{RBM}}
}

