% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{darchBench}
\alias{darchBench}
\title{Benchmarking wrapper for \code{darch}}
\usage{
darchBench(..., bench.times = 1, bench.save = F,
  bench.path = "./darch.benchmark", bench.continue = T, bench.delete = F,
  bench.plots = bench.save, output.capture = bench.save,
  plot.classificationErrorRange = 1)
}
\arguments{
\item{bench.times}{How many benchmark runs to perform}

\item{bench.save}{Whether to save benchmarking results to a directory}

\item{bench.path}{Path (relative or absolute) including directory where
benchmark results are saved if \code{bench.save} is true}

\item{bench.continue}{Whether the benchmark is to be continued from an
earlier run (will look for results of an earlier run in the specified
directory)}

\item{bench.delete}{Whether to delete the contents of the given directory if
\code{bench.continue} is \code{FALSE}. Caution: This will attempt to delete
ALL files in the given directory, use at your own risk!}

\item{bench.plots}{Whether to create plots for error rates and run times,
will be ignored if \code{bench.save} is \code{FALSE}}

\item{output.capture}{Whether to capture R output in \code{.Rout} files in
the given directory. This is the only way of gaining access to the R
output since the foreach loop will not print anything to the console. Will
be ignored if \code{bench.save} is \code{FALSE}}

\item{plot.classificationErrorRange}{Allows specification of the error range
for the classification error to make the plot more meaningful. A value of
\code{0.5}, for example, would limit the values on the y-axis to 50% of the
complete error range.}
}
\description{
Simple benchmarking function which wraps around the \code{\link{darch}}
function for users who can't or don't want to use the caret package for
benchmarking. This function requires the foreach package to work, and will
perform parallel benchmarks if an appropriate backend was registered
beforehand.
}

