% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{darchModelInfo}
\alias{darchModelInfo}
\title{Creates a custom caret model for \code{darch}.}
\usage{
darchModelInfo(params = NULL, grid = NULL)
}
\arguments{
\item{params}{\code{\link{data.frame}} of parameters or \code{NULL} to use a
simple default (bp.learnRate)}

\item{grid}{\code{\link{data.frame}} containing a grid of parameter
combinations or \code{NULL} to use a simple default}
}
\description{
TODO documentation, link to caret custom models page
}

