% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{darchTest}
\alias{darchTest}
\title{Test classification network.}
\usage{
darchTest(darch, newdata = NULL, targets = T)
}
\arguments{
\item{darch}{\code{\linkS4class{DArch}} instance.}

\item{newdata}{New data to use, \code{NULL} to use training data.}

\item{targets}{Labels for the \code{data}, \code{NULL} to use training
labels (only possible when \code{data} is \code{NULL} as well).}
}
\value{
Vector containing error function output and percentage of incorrect
 classifications.
}
\description{
Forward-propagate given data through the deep neural network and return
classification accuracy using the given labels.
}
\seealso{
Other darch interface functions: \code{\link{darch}},
  \code{\link{predict.DArch}}, \code{\link{print.DArch}}
}

