% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{example.cg}
\alias{example.cg}
\title{darch example for the iris data set using conjugate gradients.}
\description{
darch example for the iris data set using conjugate gradients.
}
\details{
For this example, we will use a simple three-layer DBN and conjugate
gradients without maxout or dropout for fine-tuning.
}
\seealso{
Other darch examples: \code{\link{example.iris}};
  \code{\link{example.maxout}};
  \code{\link{example.mnist}};
  \code{\link{example.xorNominal}};
  \code{\link{example.xor}}
}

