% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{example.iris}
\alias{example.iris}
\title{darch example for the iris data set.}
\description{
darch example for the iris data set.
}
\details{
For this example, we will use a simple three-layer DBN and backpropagation
without maxour or dropout for fine-tuning.
}
\seealso{
Other darch examples: \code{\link{example.maxout}};
  \code{\link{example.mnist}};
  \code{\link{example.xorNominal}};
  \code{\link{example.xor}}
}

