% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{example.mnist}
\alias{example.mnist}
\title{darch example using the MNIST database of handwritten digits.}
\description{
darch example using the MNIST database of handwritten digits.
}
\details{
A relatively small DBN (784, 100, and 10 neurons) is used to allow training
within a reasonable time. Increase the second number and/or add additional
layers to achieve a better training performance.

5 Epochs of RBM pre-training and 20 epochs of backpropagation fine-tuning
are used on 1000 training samples.
}
\seealso{
Other darch.examples: \code{\link{example.iris}};
  \code{\link{example.maxout}};
  \code{\link{example.xorNominal}};
  \code{\link{example.xor}}
}

