% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{example.xorNominal}
\alias{example.xorNominal}
\title{darch XOR example with nominal data.}
\description{
darch XOR example with nominal data.
}
\details{
This time with nominal data, to demonstrate that darch can handle such data
as well.

See the \code{\link{example.xor}} function for more details on the problem
itself.
}
\seealso{
Other darch examples: \code{\link{example.cg}};
  \code{\link{example.iris}}; \code{\link{example.maxout}};
  \code{\link{example.mnist}}; \code{\link{example.xor}}
}

