% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{exponentialLinearUnit}
\alias{exponentialLinearUnit}
\title{Exponential linear unit (ELU) function with unit derivatives.}
\usage{
exponentialLinearUnit(input, alpha = getDarchParam("darch.elu.alpha", 1, ...),
  ...)
}
\arguments{
\item{input}{Input for the activation function.}

\item{alpha}{ELU hyperparameter.}

\item{...}{Additional parameters, passed to \code{\link{getDarchParam}}.}
}
\value{
A list with the ELU activation in the first entry and
 the derivative of the activation in the second entry
}
\description{
The function calculates the activation of the units and returns a list, in
which the first entry is the exponential linear activation of the units and
the second entry is the derivative of the transfer function.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{linearUnit}},
  \code{\link{maxoutUnit}},
  \code{\link{rectifiedLinearUnit}},
  \code{\link{sigmoidUnit}}, \code{\link{softmaxUnit}},
  \code{\link{softplusUnit}}, \code{\link{tanhUnit}}
}

