% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{generateDropoutMask}
\alias{generateDropoutMask}
\title{Dropout mask generator function.}
\usage{
generateDropoutMask(length, dropoutRate)
}
\arguments{
\item{length}{length of the dropout mask}

\item{dropoutRate}{The dropout rate (i.e. percentage of zeros in the mask)}
}
\value{
Matrix containing the dropout mask
}
\description{
This function generates droput masks as a vector of given length, according
to the given dropout rate.
}
\keyword{internal}

