% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRBMs.R
\docType{methods}
\name{generateRBMs}
\alias{generateRBMs}
\alias{generateRBMs,DArch-method}
\title{Generates the RBMs for the pre-training.}
\usage{
generateRBMs(darch)

\S4method{generateRBMs}{DArch}(darch)
}
\arguments{
\item{darch}{An instance of the class \code{\link{DArch}} for which RBMs are
to be generated.}
}
\value{
The DArch object with the generated RBMs
}
\description{
Used the layer sizes from the DArch object to create the RBM objects for the
pre-training.
}
\seealso{
\code{\link{DArch}}
         \code{\link{RBM}}
}
\keyword{internal}

