% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeightsFunctions.R
\name{generateWeightsGlorotNormal}
\alias{generateWeightsGlorotNormal}
\title{Glorot normal weight initialization}
\usage{
generateWeightsGlorotNormal(numUnits1, numUnits2, ...)
}
\arguments{
\item{numUnits1}{Number of units in the lower layer}

\item{numUnits2}{Number of units in the upper layer}
}
\description{
This function is used to generate random weights and biases using
Glorot normal weight initialization as described in
Glorot & Bengio, AISTATS 2010.
}
\seealso{
\code{\link{Net}}

Other weight generation functions: \code{\link{generateWeightsGlorotUniform}},
  \code{\link{generateWeightsHeNormal}},
  \code{\link{generateWeightsHeUniform}},
  \code{\link{generateWeightsNormal}},
  \code{\link{generateWeightsUniform}}
}

