% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeightsFunctions.R
\name{generateWeightsGlorotUniform}
\alias{generateWeightsGlorotUniform}
\title{Glorot uniform weight initialization}
\usage{
generateWeightsGlorotUniform(numUnits1, numUnits2, ...)
}
\arguments{
\item{numUnits1}{Number of units in the lower layer.}

\item{numUnits2}{Number of units in the upper layer.}

\item{...}{Additional parameters, used for parameter resolution and passed
to \code{\link{generateWeightsUniform}}.}
}
\description{
This function is used to generate random weights and biases using
Glorot uniform weight initialization as described in
Glorot & Bengio, AISTATS 2010.
}
\seealso{
\code{\link{Net}}

Other weight generation functions: \code{\link{generateWeightsGlorotNormal}},
  \code{\link{generateWeightsHeNormal}},
  \code{\link{generateWeightsHeUniform}},
  \code{\link{generateWeightsNormal}},
  \code{\link{generateWeightsUniform}}
}

