% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeightsFunctions.R
\name{generateWeightsHeUniform}
\alias{generateWeightsHeUniform}
\title{He uniform weight initialization}
\usage{
generateWeightsHeUniform(numUnits1, numUnits2, ...)
}
\arguments{
\item{numUnits1}{Number of units in the lower layer}

\item{numUnits2}{Number of units in the upper layer}
}
\description{
This function is used to generate random weights and biases using
He uniform weight initialization as described in
He et al., \url{http://arxiv.org/abs/1502.01852}.
}
\seealso{
\code{\link{Net}}

Other weight generation functions: \code{\link{generateWeightsGlorotNormal}},
  \code{\link{generateWeightsGlorotUniform}},
  \code{\link{generateWeightsHeNormal}},
  \code{\link{generateWeightsNormal}},
  \code{\link{generateWeightsUniform}}
}

