% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateWeights.R
\name{generateWeightsRnorm}
\alias{generateWeightsRnorm}
\title{Generates a weight matrix using \link{rnorm}.}
\usage{
generateWeightsRnorm(numUnits1, numUnits2, weights.mean = 0,
  weights.sd = 0.025)
}
\arguments{
\item{numUnits1}{Number of units in the lower layer.}

\item{numUnits2}{Number of units in the upper layer.}

\item{weights.mean}{\code{mean} parameter to the \link{rnorm} function}

\item{weights.sd}{\code{sd} parameter to the \link{rnorm} function}
}
\description{
This function is the standard method for generating weights for instances of
\code{\link{Net}}. It uses \link{rnorm} to do so.
}
\seealso{
\code{\link{Net}}

Other weight generation functions: \code{\link{generateWeightsRunif}}
}

