% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getBatchSize}
\alias{getBatchSize}
\alias{getBatchSize,Net-method}
\title{Returns the batch size of the \code{\link{Net}}.}
\usage{
getBatchSize(net)

\S4method{getBatchSize}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the batch size of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

