% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getCancel,DArch-method}
\alias{getCancel,DArch-method}
\title{Returns the cancel value}
\usage{
\S4method{getCancel}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the cancel value
}
\seealso{
\link{getCancel}
}

