\docType{methods}
\name{getCancel}
\alias{getCancel}
\alias{getCancel,DArch-method}
\title{Returns the cancel value.}
\usage{
getCancel(darch)

\S4method{getCancel}{DArch}(darch)
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
Returns the cancel value.
}
\seealso{
\code{\link{DArch}}
}

