% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getCancel}
\alias{getCancel}
\title{Returns the cancel value}
\usage{
getCancel(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Cancel value.
}
\description{
Returns the cancel value
}
\seealso{
\code{\linkS4class{DArch}}
}

