% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getCancelMessage,DArch-method}
\alias{getCancelMessage,DArch-method}
\title{Returns the cancel message}
\usage{
\S4method{getCancelMessage}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the cancel message
}
\seealso{
\link{getCancelMessage}
}

