\docType{methods}
\name{getCancelMessage}
\alias{getCancelMessage}
\alias{getCancelMessage,DArch-method}
\title{Returns the cancel message.}
\usage{
getCancelMessage(darch)

\S4method{getCancelMessage}{DArch}(darch)
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
Returns the message, why the learing is canceled. If the
message is not set, the default value "no reason specified"
will be returned.
}
\seealso{
\code{\link{DArch}}
}

