% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getCancelMessage}
\alias{getCancelMessage}
\title{Returns the cancel message}
\usage{
getCancelMessage(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Cancel message.
}
\description{
Returns the message, why the learning is canceled. If the message is not set,
the default value "no reason specified" will be returned.
}
\seealso{
\code{\linkS4class{DArch}}
}

