% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getDropoutHiddenLayers,DArch-method}
\alias{getDropoutHiddenLayers,DArch-method}
\title{Returns the dropout rate for the hidden layers}
\usage{
\S4method{getDropoutHiddenLayers}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the dropout rate for the hidden layers
}
\seealso{
\link{getDropoutHiddenLayers}
}

