% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getDropoutHiddenLayers}
\alias{getDropoutHiddenLayers}
\title{Returns the dropout rate for the hidden layers}
\usage{
getDropoutHiddenLayers(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Dropout rate for the hidden layers.
}
\description{
Returns the dropout rate for the hidden layers
}
\seealso{
\code{\linkS4class{DArch}}
}

