% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getDropoutInputLayer,DArch-method}
\alias{getDropoutInputLayer,DArch-method}
\title{Returns the dropout rate for the input layer}
\usage{
\S4method{getDropoutInputLayer}{DArch}(darch)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the dropout rate for the input layer
}
\seealso{
\link{getDropoutInputLayer}
}

