% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getDropoutInputLayer}
\alias{getDropoutInputLayer}
\title{Returns the dropout rate for the input layer}
\usage{
getDropoutInputLayer(darch)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Dropout rate for the input layer.
}
\description{
Returns the dropout rate for the input layer
}
\seealso{
\code{\linkS4class{DArch}}
}

