% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getDropoutMask,DArch-method}
\alias{getDropoutMask,DArch-method}
\title{Returns the dropout mask for the given layer}
\usage{
\S4method{getDropoutMask}{DArch}(darch, i)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{i}{Layer index or 0 for input dropout mask.}
}
\description{
Returns the dropout mask for the given layer
}
\seealso{
\link{getDropoutMask}
}

