% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getDropoutMask}
\alias{getDropoutMask}
\title{Returns the dropout mask for the given layer}
\usage{
getDropoutMask(darch, i)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{i}{Layer index or 0 for input dropout mask.}
}
\value{
Dropout mask for the given layer.
}
\description{
The dropout mask is applied to the weights between layer i and i+1, for 0 < i
< numLayers. For i = 0, the dropout mask for the input layer is returned,
which will be applied to the initial input data.
}
\seealso{
\code{\linkS4class{DArch}}
}
\keyword{internal}

