% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getDropoutMasks,DArch-method}
\alias{getDropoutMasks,DArch-method}
\title{Returns the dropout masks}
\usage{
\S4method{getDropoutMasks}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the dropout masks
}
\seealso{
\link{getDropoutMasks}
}

