% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getDropoutMasks}
\alias{getDropoutMasks}
\title{Returns the dropout masks}
\usage{
getDropoutMasks(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
List of dropout masks
}
\description{
Only available during fine-tuning, returns an empty list otherwise. Unlike
\link{getDropoutMask}, this list is 1-based, starting with the
dropout mask for the input layer at index 1. So \code{getDropoutMask(darch,
0)} returns the same as \code{getDropoutMasks(darch)[[1]]}.
}
\seealso{
\code{\linkS4class{DArch}}
}

