% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getDropoutOneMaskPerEpoch,DArch-method}
\alias{getDropoutOneMaskPerEpoch,DArch-method}
\title{Return the dropout usage}
\usage{
\S4method{getDropoutOneMaskPerEpoch}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Return the dropout usage
}
\seealso{
\link{getDropoutOneMaskPerEpoch}
}

