% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getDropoutOneMaskPerEpoch}
\alias{getDropoutOneMaskPerEpoch}
\title{Return the dropout usage}
\usage{
getDropoutOneMaskPerEpoch(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Dropout usage.
}
\description{
Return whether the same dropout mask should be used for all batches of an
epoch, or a new set of masks should be generated for each batch (default).
}
\seealso{
\code{\linkS4class{DArch}}
}

