% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getEpochs,Net-method}
\alias{getEpochs,Net-method}
\title{Returns the number of epochs the \code{\linkS4class{Net}} was trained for}
\usage{
\S4method{getEpochs}{Net}(net)
}
\arguments{
\item{net}{An instance of the class \code{\linkS4class{Net}}.}
}
\description{
Returns the number of epochs the \code{\linkS4class{Net}} was trained for
}
\seealso{
\link{getEpochs}
}

