% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\name{getEpochs}
\alias{getEpochs}
\title{Returns the number of epochs the \code{\linkS4class{Net}} was trained for}
\usage{
getEpochs(net)
}
\arguments{
\item{net}{An instance of the class \code{\linkS4class{Net}}.}
}
\description{
Returns the number of epochs the \code{\linkS4class{Net}} was trained for
}

