% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getExecOutput,DArch-method}
\alias{getExecOutput,DArch-method}
\title{Returns the execution output of the layer from the \code{\linkS4class{DArch}}
object}
\usage{
\S4method{getExecOutput}{DArch}(darch, index = NULL)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{index}{The index of the layer.}
}
\description{
Returns the execution output of the layer from the \code{\linkS4class{DArch}}
object
}
\seealso{
\link{getExecOutput}
}

