% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getExecOutput}
\alias{getExecOutput}
\title{Returns the execution output of the layer from the \code{\linkS4class{DArch}}
object}
\usage{
getExecOutput(darch, index = 1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{index}{The index of the layer.}
}
\value{
The output array of the layer.
}
\description{
Returns the execution output of the layer by the given \code{index}. If the
index is not set, the output of the last layer will be returned.
}
\seealso{
\code{\linkS4class{DArch}}
}

