% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getExecOutputs,DArch-method}
\alias{getExecOutputs,DArch-method}
\title{Returns the execution output list of the \code{\linkS4class{DArch}} object}
\usage{
\S4method{getExecOutputs}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}}
}
\description{
Returns the execution output list of the \code{\linkS4class{DArch}} object
}
\seealso{
\link{getExecOutputs}
}

