\docType{methods}
\name{getExecOutputs}
\alias{getExecOutputs}
\alias{getExecOutputs,DArch-method}
\title{Returns the execution output list of the
\code{\link{DArch}} object}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
  Returns the execution output of the \code{\link{DArch}}
  object. The list contains all outputs of every layer in
  the network.
}
\seealso{
  \code{\link{DArch}}
}

