% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getExecOutputs}
\alias{getExecOutputs}
\title{Returns the execution output list of the \code{\linkS4class{DArch}} object}
\usage{
getExecOutputs(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}}
}
\value{
A list with all output of every layer.
}
\description{
Returns the execution output of the \code{\linkS4class{DArch}} object. The
list contains all outputs of every layer in the network.
}
\seealso{
\code{\linkS4class{DArch}}
}

