% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getExecuteFunction,DArch-method}
\alias{getExecuteFunction,DArch-method}
\title{Returns the function for the execution of the \code{\linkS4class{DArch}}
network}
\usage{
\S4method{getExecuteFunction}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the function for the execution of the \code{\linkS4class{DArch}}
network
}
\seealso{
\link{getExecuteFunction}
}

