% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getExecuteFunction}
\alias{getExecuteFunction}
\title{Returns the function for the execution of the \code{\linkS4class{DArch}}
network}
\usage{
getExecuteFunction(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Returns the execution, aka forward propagation, function.
}
\description{
Returns the function for the execution of the \code{\linkS4class{DArch}}
network
}
\seealso{
\code{\linkS4class{DArch}}
}

