\docType{methods}
\name{getFinalMomentum}
\alias{getFinalMomentum}
\alias{getFinalMomentum,Net-method}
\title{Returns the final momentum of the \code{\link{Net}}.}
\usage{
getFinalMomentum(net)

\S4method{getFinalMomentum}{Net}(net)
}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the final momentum of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

