% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getFineTuneFunction,DArch-method}
\alias{getFineTuneFunction,DArch-method}
\title{Returns the fine tune function for the \code{\linkS4class{DArch}} object}
\usage{
\S4method{getFineTuneFunction}{DArch}(darch)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the fine tune function for the \code{\linkS4class{DArch}} object
}
\seealso{
\link{getFineTuneFunction}
}

