\docType{methods}
\name{getHiddenBiasesInc}
\alias{getHiddenBiasesInc}
\alias{getHiddenBiasesInc,RBM-method}
\title{Returns the update value for the biases of the hidden units.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The update value for the biases of the hidden units.
}
\description{
  Returns the update value for the biases of the hidden
  units.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

