\docType{methods}
\name{getHiddenUnitStates}
\alias{getHiddenUnitStates}
\alias{getHiddenUnitStates,RBM-method}
\title{Returns a list with the states of the hidden units.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The states of the hidden units.
}
\description{
  Returns a list with the states of the hidden units.
}
\seealso{
  \code{\link{RBM}}
}

