% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getInitialMomentum,Net-method}
\alias{getInitialMomentum,Net-method}
\title{Returns the momentum of the \code{\link{Net}}}
\usage{
\S4method{getInitialMomentum}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the momentum of the \code{\link{Net}}
}
\seealso{
\link{getInitialMomentum}
}

