% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\name{getInitialMomentum}
\alias{getInitialMomentum}
\title{Returns the momentum of the \code{\link{Net}}}
\usage{
getInitialMomentum(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the momentum of the \code{\link{Net}}
}
\seealso{
\code{\link{Net}}
}

