\docType{methods}
\name{getLayer}
\alias{getLayer}
\alias{getLayer,DArch-method}
\title{Returns a layer from the \code{\link{DArch}} object.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{index}{The index of the layer.Default is 1.}
}
\description{
  The function returns the layer with the given
  \code{index} which contains all weights, functions for
  the neurons, and possible additional parameters for the
  training.
}
\seealso{
  \code{\link{DArch}}
}

