% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getLayer}
\alias{getLayer}
\title{Returns a layer from the \code{\linkS4class{DArch}} object}
\usage{
getLayer(darch, index = 1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{index}{The index of the layer. Default is 1.}
}
\value{
A layers from the \code{\linkS4class{DArch}} object.
}
\description{
The function returns the layer with the given \code{index} which contains all
weights, functions for the neurons, and possible additional parameters for
the training.
}
\seealso{
\code{\linkS4class{DArch}}
}

