% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getLayerField,DArch-method}
\alias{getLayerField,DArch-method}
\title{Returns the field of a layer from the \code{\linkS4class{DArch}} object}
\usage{
\S4method{getLayerField}{DArch}(darch, layerIndex = 1, fieldIndex = 3)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{layerIndex}{The index of the layer.}

\item{fieldIndex}{The index of the field in the layer.}
}
\description{
Returns the field of a layer from the \code{\linkS4class{DArch}} object
}
\seealso{
\link{getLayerField}
}

