\docType{methods}
\name{getLayerField}
\alias{getLayerField}
\alias{getLayerField,DArch-method}
\title{Returns the field of a layer from the
\code{\link{DArch}} object.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{layerIndex}{The index of the layer.}

  \item{fieldIndex}{The index of the field in the layer.}
}
\description{
  The function returns the field given by the
  \code{fieldIndex} of the layer given by the
  \code{layerIndex}.
}
\seealso{
  \code{\link{DArch}}
}

