% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getLayerFunction}
\alias{getLayerFunction}
\title{Returns the neuron function of a layer from the \code{\linkS4class{DArch}}
object}
\usage{
getLayerFunction(darch, index = 1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{index}{The index of the layer.Default is 1.}
}
\value{
A neuron function.
}
\description{
The function returns the neuron function of the layer with the given
\code{index}.
}
\seealso{
\code{\linkS4class{DArch}} \code{\link{sigmoidUnit}},
  \code{\link{binSigmoidUnit}}, \code{\link{sigmoidUnitDerivative}},
  \code{\link{linearUnitDerivative}}, \code{\link{softmaxUnit}},
  \code{\link{softmaxUnitDerivative}}
}

