% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getLayers,DArch-method}
\alias{getLayers,DArch-method}
\title{Returns the a list of layers from the \code{\linkS4class{DArch}} object.}
\usage{
\S4method{getLayers}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the a list of layers from the \code{\linkS4class{DArch}} object.
}
\seealso{
\link{getLayers}
}

