\docType{methods}
\name{getLayers}
\alias{getLayers}
\alias{getLayers,DArch-method}
\title{Returns the a list of layers from the \code{\link{DArch}}
object.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
  The function returns the layers list which contains all
  weights, functions for the neurons, and possible
  additional parameters for the training.
}
\seealso{
  \code{\link{DArch}}
}

