% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getLayers}
\alias{getLayers}
\title{Returns the a list of layers from the \code{\linkS4class{DArch}} object}
\usage{
getLayers(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
A list of layers from the \code{\linkS4class{DArch}} object.
}
\description{
The function returns the layers list which contains all weights, functions
for the neurons, and possible additional parameters for the training.
}
\seealso{
\code{\linkS4class{DArch}}
}

