% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getLayerWeights,DArch-method}
\alias{getLayerWeights,DArch-method}
\title{Returns the weights of a layer from the \code{\linkS4class{DArch}} object}
\usage{
\S4method{getLayerWeights}{DArch}(darch, index = 1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{index}{The index of the layer.Default is 1.}
}
\description{
Returns the weights of a layer from the \code{\linkS4class{DArch}} object
}
\seealso{
\link{getLayerWeights}
}

