% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getLayerWeights}
\alias{getLayerWeights}
\title{Returns the weights of a layer from the \code{\linkS4class{DArch}} object}
\usage{
getLayerWeights(darch,index=1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{index}{The index of the layer.Default is 1.}
}
\value{
A weight matrix.
}
\description{
The function returns the weights of the layer with the given \code{index}.
}
\seealso{
\code{\linkS4class{DArch}}
}

