% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getLearnRateBiases,DArch-method}
\alias{getLearnRateBiases,DArch-method}
\title{Returns the learning rate for the bias weights of the
\code{\linkS4class{DArch}} object}
\usage{
\S4method{getLearnRateBiases}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns the learning rate for the bias weights of the
\code{\linkS4class{DArch}} object
}
\seealso{
\link{getLearnRateBiases}
}

