\docType{methods}
\name{getLearnRateBiases}
\alias{getLearnRateBiases}
\alias{getLearnRateBiases,DArch-method}
\title{Returns the learning rate for the bias weigths of the \code{\link{DArch}}
object.}
\usage{
getLearnRateBiases(darch)

\S4method{getLearnRateBiases}{DArch}(darch)
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
Returns the learning rate for the bias weigths of the
\code{\link{DArch}} object.
}
\seealso{
\code{\link{DArch}}
}

