% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getLearnRateBiases}
\alias{getLearnRateBiases}
\title{Returns the learning rate for the bias weights of the
\code{\linkS4class{DArch}} object}
\usage{
getLearnRateBiases(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
Learning rate of the biases.
}
\description{
Returns the learning rate for the bias weights of the
\code{\linkS4class{DArch}} object
}
\seealso{
\code{\linkS4class{DArch}}
}

