\docType{methods}
\name{getLearnRateBiasHidden}
\alias{getLearnRateBiasHidden}
\alias{getLearnRateBiasHidden,RBM-method}
\title{Returns the learning rate for the hidden biases.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The learning rate for the hidden biases
}
\description{
  Returns the learning rate for the hidden biases.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

