\docType{methods}
\name{getLearnRateBiasVisible}
\alias{getLearnRateBiasVisible}
\alias{getLearnRateBiasVisible,RBM-method}
\title{Returns the learning rate for the visible biases.}
\usage{
getLearnRateBiasVisible(rbm)

\S4method{getLearnRateBiasVisible}{RBM}(rbm)
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The learning rate for the visible biases
}
\description{
Returns the learning rate for the visible biases.
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

