\docType{methods}
\name{getLogger}
\alias{getLogger}
\alias{getLogger,Net-method}
\title{Returns the logger of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns the logger of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}} and \code{\link[log4r]{create.logger}}
}

