\docType{methods}
\name{getMomentum}
\alias{getMomentum}
\alias{getMomentum,Net-method}
\title{Returns the momentum of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns the momentum of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

