% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.Getter.R
\name{getMomentum}
\alias{getMomentum}
\title{Returns the current momentum of the \code{\link{Net}}.}
\usage{
getMomentum(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the current momentum of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}
\keyword{internal}

