% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getMomentumSwitch}
\alias{getMomentumSwitch}
\alias{getMomentumSwitch,Net-method}
\title{Returns the momentum switch of the \code{\link{Net}}.}
\usage{
getMomentumSwitch(net)

\S4method{getMomentumSwitch}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the momentum switch of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

