% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getNormalizeWeights,Net-method}
\alias{getNormalizeWeights,Net-method}
\title{Returns whether weight normalization is active}
\usage{
\S4method{getNormalizeWeights}{Net}(net)
}
\arguments{
\item{net}{An instance of the class \code{\linkS4class{Net}}.}
}
\description{
Returns whether weight normalization is active
}
\seealso{
\link{getNormalizeWeights}
}

