% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\name{getNormalizeWeights}
\alias{getNormalizeWeights}
\title{Returns whether weight normalization is active}
\usage{
getNormalizeWeights(net)
}
\arguments{
\item{net}{An instance of the class \code{\linkS4class{Net}}.}
}
\description{
Returns whether weight normalization is active
}

