% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getNumHidden}
\alias{getNumHidden}
\alias{getNumHidden,RBM-method}
\title{Returns the number of hidden units of the \code{\link{RBM}}}
\usage{
getNumHidden(rbm)

\S4method{getNumHidden}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The number of hidden units of the \code{\link{RBM}}
}
\description{
Returns the number of hidden units of the \code{\link{RBM}}
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

