% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getOutput}
\alias{getOutput}
\alias{getOutput,RBM-method}
\title{Returns the output of the \code{\link{RBM}}}
\usage{
getOutput(rbm)

\S4method{getOutput}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The output of the \code{\link{RBM}}
}
\description{
Returns the output of the \code{\link{RBM}}
}
\seealso{
\code{\link{RBM}}
}

