\docType{methods}
\name{getPosPhaseData}
\alias{getPosPhaseData}
\alias{getPosPhaseData,RBM-method}
\title{Returns the data for the positive phase.}
\usage{
getPosPhaseData(rbm)

\S4method{getPosPhaseData}{RBM}(rbm)
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The data for the positive phase.
}
\description{
Returns the data for the positive phase.
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

