% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getRBMList,DArch-method}
\alias{getRBMList,DArch-method}
\title{Returns a list of \code{\link{RBM}}s of the \code{\linkS4class{DArch}}
object}
\usage{
\S4method{getRBMList}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\description{
Returns a list of \code{\link{RBM}}s of the \code{\linkS4class{DArch}}
object
}
\seealso{
\link{getRBMList}
}

