\docType{methods}
\name{getRBMList}
\alias{getRBMList}
\alias{getRBMList,DArch-method}
\title{Returns a list of \code{\link{RBM}}s of the
\code{\link{DArch}} object.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
  This function returns a list of \code{\link{RBM}}s of the
  \code{\link{DArch}} which are needed for the pre training
  of the network.
}
\seealso{
  \code{\link{DArch}}
}

