% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getRBMList}
\alias{getRBMList}
\title{Returns a list of \code{\link{RBM}}s of the \code{\linkS4class{DArch}}
object}
\usage{
getRBMList(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}
}
\value{
A list of all code{\link{RBM}}s of the \code{\linkS4class{DArch}}
  object.
}
\description{
This function returns a list of \code{\link{RBM}}s of the
\code{\linkS4class{DArch}} which are needed for the pre training of the
network.
}
\seealso{
\code{\linkS4class{DArch}}
}

