% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getStats}
\alias{getStats}
\alias{getStats,Net-method}
\title{Returns the list of statistics for the network}
\usage{
getStats(net)

\S4method{getStats}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
The list of statistics can contain values about errors, miss
classifications and other useful things from the pre-training or fine-tuning
 of a deep architecture.
}
\seealso{
\code{\link{Net}}
}

