% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getVisibleBiasesInc}
\alias{getVisibleBiasesInc}
\alias{getVisibleBiasesInc,RBM-method}
\title{Returns the update value for the biases of the visible units.}
\usage{
getVisibleBiasesInc(rbm)

\S4method{getVisibleBiasesInc}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The update value for the biases of the visible units.
}
\description{
Returns the update value for the biases of the visible units.
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

