\docType{methods}
\name{getWeightCost}
\alias{getWeightCost}
\alias{getWeightCost,RBM-method}
\title{Returns the weight cost for the training}
\usage{
getWeightCost(rbm)

\S4method{getWeightCost}{RBM}(rbm)
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The weight cost for the training
}
\description{
Returns the weight cost for the training
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

