% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\docType{methods}
\name{getWeightUpdateFunction,DArch-method}
\alias{getWeightUpdateFunction,DArch-method}
\title{Returns the weight update function}
\usage{
\S4method{getWeightUpdateFunction}{DArch}(darch, layerIndex = 1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{layerIndex}{The index of the layer. Default is 1.}
}
\description{
Returns the weight update function
}
\seealso{
\link{getWeightUpdateFunction}
}

