% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Getter.R
\name{getWeightUpdateFunction}
\alias{getWeightUpdateFunction}
\title{Return the weight update function.}
\usage{
getWeightUpdateFunction(darch, layerIndex = 1)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{layerIndex}{The index of the layer. Default is 1.}
}
\value{
A weight update function.
}
\description{
Returns the weight update function of the given layer.
}
\seealso{
\code{\linkS4class{DArch}}, \link{setWeightUpdateFunction},
 \link{weightDecayWeightUpdate}
}

