% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{incrementEpochs,Net-method}
\alias{incrementEpochs,Net-method}
\title{Increment the number of epochs this \code{\linkS4class{Net}} has been trained
for}
\usage{
\S4method{incrementEpochs}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Increment the number of epochs this \code{\linkS4class{Net}} has been trained
for
}
\seealso{
\link{incrementEpochs}
}

