% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\name{incrementEpochs}
\alias{incrementEpochs}
\title{Increment the number of epochs this \code{\linkS4class{Net}} has been trained
for}
\usage{
incrementEpochs(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Increment the number of epochs this \code{\linkS4class{Net}} has been trained
for
}

