% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{linearUnit}
\alias{linearUnit}
\title{Linear unit function.}
\usage{
linearUnit(data, weights)
}
\arguments{
\item{data}{The data matrix for the calculation}

\item{weights}{The weight and bias matrix for the calculation}
}
\value{
A list with the linear activation of the unit in the first entry.
}
\description{
The function calculates the activation of the units and returns it.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{binSigmoidUnit}};
  \code{\link{linearUnitDerivative}};
  \code{\link{maxoutUnitDerivative}};
  \code{\link{sigmoidUnitDerivative}};
  \code{\link{sigmoidUnit}};
  \code{\link{softmaxUnitDerivative}};
  \code{\link{softmaxUnit}};
  \code{\link{tanSigmoidUnitDerivative}};
  \code{\link{tanSigmoidUnit}}
}

