\docType{methods}
\name{linearUnitFunc}
\alias{linearUnitFunc}
\title{BEARBEITEN}
\usage{
  linearUnitFunc(rbm, dataList, biases, weights, runParams)
}
\arguments{
  \item{rbm}{BEARBEITEN}

  \item{dataList}{BEARBEITEN}

  \item{biases}{BEARBEITEN}

  \item{weights}{BEARBEITEN}

  \item{runParams}{BEARBEITEN}

  \item{...}{BEARBEITEN}
}
\value{
  The activations for the units
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{DArch}}
}

