\docType{methods}
\name{minimize}
\alias{minimize}
\title{BEARBEITEN}
\usage{
  minimize(X, f, length, ...)
}
\arguments{
  \item{X}{BEARBEITEN}

  \item{f}{BEARBEITEN}

  \item{length}{BEARBEITEN}

  \item{...}{BEARBEITEN}
}
\value{
  BEARBEITEN
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{DArch}}
}

