\docType{methods}
\name{minimizeAutoencoder}
\alias{minimizeAutoencoder}
\title{BEARBEITEN}
\usage{
  minimizeAutoencoder(darch, trainData, targetData, epoch,
    length)
}
\arguments{
  \item{darch}{BEARBEITEN}

  \item{trainData}{BEARBEITEN}

  \item{targetData}{BEARBEITEN}

  \item{epoch}{BEARBEITEN}

  \item{...}{BEARBEITEN}
}
\value{
  BEARBEITEN
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{DArch}}
}

