% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.R
\docType{class}
\name{Net-class}
\alias{Net}
\alias{Net-class}
\title{Abstract class for neural networks.}
\description{
This is a abstract class for neural networks. It provides some
functionalities used in more than one network type.
}
\section{Slots}{

\describe{
\item{\code{batchSize}}{Object of class \code{"numeric"}. The batch size
for the training and test data during the learning.}

\item{\code{errorFunction}}{Object of class \code{"function"}. Function for error
calculation.}

\item{\code{ff}}{Object of class \code{"logical"}. Indicates if the package
\code{\link[ff]{ff}} is used to save the network data.}

\item{\code{genWeightFunction}}{Object of class \code{"function"}. A function to
generate a randomly initialised weight matrix.}

\item{\code{normalizeWeights}}{Logical indicating whether to normalize the weights.}

\item{\code{learnRateWeights}}{Learn rate.}

\item{\code{initialMomentum}}{Initial momentum.}

\item{\code{finalMomentum}}{Final momentum.}

\item{\code{momentumSwitch}}{Epoch at which to switch from \code{initialMomentum} to
\code{finalMomentum}.}

\item{\code{epochs}}{Number of epochs.}

\item{\code{stats}}{Training statistics.}
}}
\author{
Martin Drees
}
\seealso{
\code{\linkS4class{DArch}}, \code{\linkS4class{RBM}}
}

