% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRBM.R
\name{newRBM}
\alias{newRBM}
\title{Constructor function for RBM object.}
\usage{
newRBM(numVisible, numHidden, batchSize, genWeightFunc, ...)
}
\arguments{
\item{numVisible}{Number of visible units.}

\item{numHidden}{Number of hidden units.}

\item{batchSize}{Size of the batches.}

\item{genWeightFunc}{The function for generating the weight matrices}
}
\value{
The new RBM object
}
\description{
Constructor function for RBM object.
}

