% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.DArch}
\alias{plot.DArch}
\title{Plot \code{\link{DArch}} statistics or structure.}
\usage{
\method{plot}{DArch}(x, y = NULL, ..., type = "errorRaw")
}
\arguments{
\item{x}{\code{\link{DArch}} instance}

\item{y}{ignored}

\item{type}{Which type of plot to create, one of \code{errorRaw},
\code{errorClass}, \code{time}, or \code{net}.}
}
\description{
This function provides different plots depending on the \code{type}
parameter:
}
\details{
\itemize{
 \item errorRaw. Prints the raw network error (e.g. MSE), this is the default
 \item errorClass. Prints the classification error
 \item time. Prints the times needed for each epoch
 \item net. Calls \code{\link{NeuralNetTools::plotnet}} to plot the network
}
}

