% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{predict.DArch}
\alias{predict.DArch}
\alias{predict.darch}
\title{Forward-propagate data.}
\usage{
\method{predict}{DArch}(object, ..., newdata = NULL, type = "raw")
}
\arguments{
\item{object}{\code{\linkS4class{DArch}} instance}

\item{...}{Further parameters, not used.}

\item{newdata}{New data to predict, \code{NULL} to return latest network
output}

\item{type}{Output type, one of: \code{raw}, \code{bin}, \code{class}.
\code{raw} returns the network output (as is, or with scaling reversed, if
the input data were scaled), \code{bin} returns \code{1} for every network
output \code{>0.5}, \code{0} otherwise, and \code{class} returns \code{1}
for the output unit with the highest activation, otherwise \code{0}.
Additionally, when using \code{class}, class labels are returned when
available.}
}
\value{
Vector or matrix of networks outputs, output type depending on the
  \code{type} parameter
}
\description{
Forward-propagate given data through the deep neural network.
}
\seealso{
Other darch interface functions: \code{\link{darch.DataSet}};
  \code{\link{darch.default}}; \code{\link{darch.formula}};
  \code{\link{darch}}; \code{\link{print.DArch}},
  \code{\link{print.darch}}
}

