% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darch.Learn.R
\docType{methods}
\name{preTrainDArch,DArch-method}
\alias{preTrainDArch,DArch-method}
\title{Pre-trains a \code{\linkS4class{DArch}} network}
\usage{
\S4method{preTrainDArch}{DArch}(darch, dataSet, dataSetValid = NULL,
  numEpochs = 1, numCD = 1, lastLayer = 0, isClass = F,
  consecutive = T, ...)
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{dataSet}{\code{\linkS4class{DataSet}} to be used for training.}

\item{dataSetValid}{\code{\linkS4class{DataSet}} to be used for validation.}

\item{numEpochs}{The number of epochs}

\item{numCD}{The number of CD iterations}

\item{lastLayer}{Numeric indicating after which layer to stop training.}

\item{isClass}{Whether to test pre-trained networks against target data}

\item{consecutive}{Whether to train RBMs consecutively instead of each one
epoch at a time.}

\item{...}{Additional parameters for the function \code{\link{trainRBM}}}
}
\description{
Pre-trains a \code{\linkS4class{DArch}} network
}
\seealso{
\link{preTrainDArch}
}
\keyword{internal}

