% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{print.DArch}
\alias{print.DArch}
\alias{print.darch}
\title{Print \linkS4class{DArch} details.}
\usage{
\method{print}{DArch}(darch)
}
\arguments{
\item{darch}{\code{\linkS4class{DArch}} instance}
}
\description{
Print verbose information about a \linkS4class{DArch} instance.
}

