\docType{methods}
\name{pseudoInverse}
\alias{pseudoInverse}
\title{BEARBEITEN}
\usage{
  pseudoInverse(darch, trainData, targetData, epoch, ...)
}
\arguments{
  \item{darch}{BEARBEITEN}

  \item{trainData}{BEARBEITEN}

  \item{targetData}{BEARBEITEN}

  \item{epoch}{BEARBEITEN}

  \item{...}{BEARBEITEN}
}
\value{
  BEARBEITEN
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{DArch}}
}

