% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{rectifiedLinearUnitDerivative}
\alias{rectifiedLinearUnitDerivative}
\title{Rectified linear unit function with unit derivatives.}
\usage{
rectifiedLinearUnitDerivative(input, ...)
}
\arguments{
\item{input}{Input for the activation function.}
}
\value{
A list with the rectified linear activation in the first entry and
 the derivative of the activation in the second entry
}
\description{
The function calculates the activation of the units and returns a list, in
which the first entry is the rectified linear activation of the units and
the second entry is the derivative of the transfer function.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{linearUnitDerivative}};
  \code{\link{maxoutUnitDerivative}};
  \code{\link{sigmoidUnitDerivative}};
  \code{\link{softmaxUnitDerivative}};
  \code{\link{tanSigmoidUnitDerivative}}
}

