\docType{methods}
\name{removeLayerField}
\alias{removeLayerField}
\alias{removeLayerField,DArch-method}
\title{Removes a layer from the \code{\link{DArch}} object}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\value{
  The \code{\link{DArch}} object without the layer.
}
\description{
  This function removes the layer with the given index from
  the \code{\link{DArch}} object.
}
\seealso{
  \code{\link{DArch}}
}

