% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Reset.R
\docType{methods}
\name{resetRBM}
\alias{resetRBM}
\alias{resetRBM,RBM-method}
\title{Resets the weights and biases of the \code{\link{RBM}} object}
\usage{
resetRBM(rbm)

\S4method{resetRBM}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\description{
This function resets the weights and biases of the \code{\link{RBM}} object.
}
\seealso{
\code{\link{RBM}}
}

