% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm.Reset.R
\name{resetRBM}
\alias{resetRBM}
\title{Resets the weights and biases of the \code{\link{RBM}} object}
\usage{
resetRBM(rbm, ...)
}
\arguments{
\item{rbm}{An instance of class \code{\link{RBM}}.}

\item{...}{Additional arguments.}
}
\description{
This function resets the weights and biases of the \code{\link{RBM}} object.
}
\seealso{
\code{\link{RBM}}
}
\keyword{internal}

