% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDArch.R
\name{runDArch}
\alias{runDArch}
\title{Execute the darch}
\usage{
runDArch(darch, data, outputLayer = 0, matMult = getDarchParam("matMult",
  `\%*\%`, darch = darch))
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{data}{The input data to execute the darch on.}

\item{outputLayer}{The output of which layer is to be returned, absolute
number or offset from the last layer.}

\item{matMult}{Function to use for matrix multiplication.}
}
\value{
The DArch object with the calculated outputs
}
\description{
Runs the darch in a feed forward manner and saves the 
generated outputs for every layer in the list
\code{executeOutput} from the darch.
}
\seealso{
\code{\link{DArch}}

Other darch execute functions: \code{\link{runDArchDropout}}
}

