\docType{methods}
\name{setBatchSize<-}
\alias{setBatchSize}
\alias{setBatchSize<-}
\alias{setBatchSize<-,Net-method}
\title{Sets the batch size of the \code{\link{Net}}.}
\usage{
setBatchSize(net) <- value
}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the batch size of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

