\docType{methods}
\name{setCancel<-}
\alias{setCancel}
\alias{setCancel<-}
\alias{setCancel<-,DArch-method}
\title{Set whether the learning shall be canceled.}
\usage{
setCancel(darch) <- value
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{Boolean value if the learning shall
  canceled}
}
\description{
Set whether the learning shall be canceled.
}
\seealso{
\code{\link{DArch}}
}

