% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setCancelMessage<-}
\alias{setCancelMessage}
\alias{setCancelMessage<-}
\alias{setCancelMessage<-,DArch-method}
\title{Sets the cancel message.}
\usage{
setCancelMessage(darch) <- value

\S4method{setCancelMessage}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{The message for the termination}
}
\description{
Sets the cancel message.
}
\seealso{
\code{\link{DArch}}
}

