% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darch.Params.R
\name{setDarchParams}
\alias{setDarchParams}
\title{Set \code{\link{DArch}} parameters}
\usage{
setDarchParams(darch, ...)
}
\arguments{
\item{darch}{\code{\link{DArch}} instance.}

\item{...}{Parameters to be set, see \code{\link{darch.default}}.}
}
\description{
Allows setting \code{\link{DArch}} parameters normally passed to the
\code{\link{darch}} interface function when not using said interface.
These parameters can also be passed to \code{\link{newDArch}}.
}
\keyword{internal}

