% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setDropoutHiddenLayers<-}
\alias{setDropoutHiddenLayers}
\alias{setDropoutHiddenLayers<-}
\alias{setDropoutHiddenLayers<-,DArch-method}
\title{Sets the dropout rate for the hidden layers.}
\usage{
setDropoutHiddenLayers(darch) <- value

\S4method{setDropoutHiddenLayers}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{Dropout rate for the hidden layers}
}
\value{
The darch with the updated dropout rate
}
\description{
Sets the dropout rate for the hidden layers.
}
\seealso{
\code{\link{DArch}}
}

