% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setDropoutInputLayer<-}
\alias{setDropoutInputLayer}
\alias{setDropoutInputLayer<-}
\alias{setDropoutInputLayer<-,DArch-method}
\title{Sets the dropout rate for the input layer.}
\usage{
setDropoutInputLayer(darch) <- value

\S4method{setDropoutInputLayer}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{Dropout rate for the input layer}
}
\value{
The darch with the updated dropout rate
}
\description{
Sets the dropout rate for the input layer.
}
\seealso{
\code{\link{DArch}}
}

