% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setDropoutMask<-}
\alias{setDropoutMask}
\alias{setDropoutMask<-}
\alias{setDropoutMask<-,DArch-method}
\title{Set the dropout mask for the given layer.}
\usage{
setDropoutMask(darch, i) <- value

\S4method{setDropoutMask}{DArch}(darch, i) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{i}{Layer index or 0 for input layer.}

\item{value}{Dropout mask for the given layer.}
}
\value{
The darch with the updated dropout mask
}
\description{
Sets the dropout mask to be applied to the weights between layer i and i+1,
for 0 < i < numLayers. For i = 0, sets the dropout mask for the input layer,
which will be applied to the initiali input data.
}
\seealso{
\code{\link{DArch}}
}

