% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darch.Setter.R
\name{setDropoutMask<-}
\alias{setDropoutMask<-}
\title{Set the dropout mask for the given layer.}
\usage{
setDropoutMask(darch, i) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{i}{Layer index or 0 for input layer.}

\item{value}{Dropout mask for the given layer.}
}
\value{
The darch with the updated dropout mask
}
\description{
Sets the dropout mask to be applied to the weights between layer i and i+1,
for 0 < i < numLayers. For i = 0, sets the dropout mask for the input layer,
which will be applied to the initiali input data.
}
\seealso{
\code{\linkS4class{DArch}}
}
\keyword{internal}

