% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setDropoutMasks<-}
\alias{setDropoutMasks}
\alias{setDropoutMasks<-}
\alias{setDropoutMasks<-,DArch-method}
\title{Set the dropout masks.}
\usage{
setDropoutMasks(darch) <- value

\S4method{setDropoutMasks}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{List of dropout masks}
}
\value{
The darch with the updated dropout masks
}
\description{
Set the dropout masks.
}
\seealso{
\code{\link{DArch}}
}

