% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setDropoutOneMaskPerEpoch<-,DArch-method}
\alias{setDropoutOneMaskPerEpoch<-,DArch-method}
\title{Set dropout mask usage}
\usage{
\S4method{setDropoutOneMaskPerEpoch}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{value}{Boolean value indicating whether to use a dropout mask for each
  epoch.}
}
\description{
Set dropout mask usage
}
\seealso{
\link{setDropoutOneMaskPerEpoch<-}
}

