% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\name{setDropoutOneMaskPerEpoch<-}
\alias{setDropoutOneMaskPerEpoch<-}
\title{Set dropout mask usage}
\usage{
setDropoutOneMaskPerEpoch(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\linkS4class{DArch}}.}

\item{value}{Boolean value indicating whether to use a dropout mask for each
  epoch.}
}
\description{
Set whether the same dropout mask should be used for all batches of an epoch,
or a new set of masks should be generated for each batch (default).
}

