% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setErrorFunction<-}
\alias{setErrorFunction}
\alias{setErrorFunction<-}
\alias{setErrorFunction<-,Net-method}
\title{Sets the error function of the \code{\link{Net}}.}
\usage{
setErrorFunction(net) <- value

\S4method{setErrorFunction}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{function}.}
}
\description{
Sets the error function of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

