\docType{methods}
\name{setExecuteFunction<-}
\alias{setExecuteFunction}
\alias{setExecuteFunction<-}
\alias{setExecuteFunction<-,DArch-method}
\title{BEARBEITEN}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{Net}}
}

