% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setFF<-}
\alias{setFF}
\alias{setFF<-}
\alias{setFF<-,Net-method}
\title{Sets if the weights are saved as ff objects}
\usage{
setFF(net) <- value

\S4method{setFF}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Boolean value which indicates if the weights are saved as ff
objects}
}
\description{
Sets if the weights are saved as ff objects
}
\seealso{
\code{\link{Net}}
}

