% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setFinalMomentum<-}
\alias{setFinalMomentum}
\alias{setFinalMomentum<-}
\alias{setFinalMomentum<-,Net-method}
\title{Sets the final momentum of the \code{\link{Net}}.}
\usage{
setFinalMomentum(net) <- value

\S4method{setFinalMomentum}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the final momentum of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

