\docType{methods}
\name{setFineTuneFunction<-}
\alias{setFineTuneFunction}
\alias{setFineTuneFunction<-}
\alias{setFineTuneFunction<-,DArch-method}
\title{Sets the fine tuning function for the network}
\usage{
setFineTuneFunction(darch)  <- value
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{The fine tuning function for the network.}
}
\description{
Sets the fine tuning function for the network
}
\seealso{
\code{\link{DArch}}
}

