\docType{methods}
\name{setHiddenBiases<-}
\alias{setHiddenBiases}
\alias{setHiddenBiases<-}
\alias{setHiddenBiases<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

