\docType{methods}
\name{setHiddenBiasesInc<-}
\alias{setHiddenBiasesInc}
\alias{setHiddenBiasesInc<-}
\alias{setHiddenBiasesInc<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

