\docType{methods}
\name{setHiddenUnitStates<-}
\alias{setHiddenUnitStates}
\alias{setHiddenUnitStates<-}
\alias{setHiddenUnitStates<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

