% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setInitialMomentum<-,Net-method}
\alias{setInitialMomentum<-,Net-method}
\title{Sets the initial momentum of the \code{\linkS4class{Net}}}
\usage{
\S4method{setInitialMomentum}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the initial momentum of the \code{\linkS4class{Net}}
}
\seealso{
\link{setInitialMomentum<-}
}

