% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\name{setInitialMomentum<-}
\alias{setInitialMomentum<-}
\title{Sets the initial momentum of the \code{\linkS4class{Net}}}
\usage{
setInitialMomentum(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the initial momentum of the \code{\linkS4class{Net}}
}

