% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setLayer<-}
\alias{setLayer}
\alias{setLayer<-}
\alias{setLayer<-,DArch-method}
\title{Sets a layer with the given index for the network}
\usage{
setLayer(darch,index) <- value

\S4method{setLayer}{DArch}(darch, index) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{index}{The index of the layer}

\item{value}{The layer for the network.}
}
\description{
Sets a layer with the given index for the network
}
\seealso{
\code{\link{DArch}}
}

