\docType{methods}
\name{setLayerField<-}
\alias{setLayerField}
\alias{setLayerField<-}
\alias{setLayerField<-,DArch-method}
\title{Sets a field in a layer.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{layerIndex}{The index of the layer}

  \item{fieldIndex}{The index of the field}

  \item{value}{The value for the layer field}
}
\value{
  The darch with the updated layer
}
\description{
  Sets the field on position \code{fieldIndex} of the layer
  given by the \code{layerIndex} to the \code{value}.
}
\seealso{
  \code{\link{DArch}}
}

