\docType{methods}
\name{setLayerFunction<-}
\alias{setLayerFunction}
\alias{setLayerFunction<-}
\alias{setLayerFunction<-,DArch-method}
\title{BEARBEITEN}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{Net}}
}

