\docType{methods}
\name{setLayers<-}
\alias{setLayers}
\alias{setLayers<-}
\alias{setLayers<-,DArch-method}
\title{Sets the layers for the network}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{The layers for the network.}
}
\description{
  Sets the layers for the network
}
\seealso{
  \code{\link{DArch}}
}

