% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\name{setLearnRate<-}
\alias{setLearnRate<-}
\title{Sets the learning rate.}
\usage{
setLearnRate(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the learning rate.
}
\seealso{
\code{\link{Net}}
}

