\docType{methods}
\name{setLearnRateBiases<-}
\alias{setLearnRateBiases}
\alias{setLearnRateBiases<-}
\alias{setLearnRateBiases<-,DArch-method}
\title{BEARBEITEN}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{Net}}
}

