\docType{methods}
\name{setLearnRateBiasHidden<-}
\alias{setLearnRateBiasHidden}
\alias{setLearnRateBiasHidden<-}
\alias{setLearnRateBiasHidden<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

