% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setLearnRateWeights<-}
\alias{setLearnRateWeights}
\alias{setLearnRateWeights<-}
\alias{setLearnRateWeights<-,Net-method}
\title{Sets the learning rate for the weights.}
\usage{
setLearnRateWeights(net) <- value

\S4method{setLearnRateWeights}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the learning rate for the weights.
}
\seealso{
\code{\link{Net}}
}

