\docType{methods}
\name{setLogger<-}
\alias{setLogger}
\alias{setLogger<-}
\alias{setLogger<-,Net-method}
\title{Sets the logger of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class
  \code{\link[log4r]{create.logger}}.}
}
\description{
  Sets the logger of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}} and \code{\link[ff]{ff}}
}

