% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{setLogLevel}
\alias{setLogLevel}
\title{Set the log level.}
\usage{
setLogLevel(value)
}
\arguments{
\item{value}{Log level, must be one of the \code{futile.logger} constants.}
}
\description{
Convenience wrapper for \code{\link[futile.logger]{flog.threshold}} with
sanity checking and current level output.
}
\details{
The log levels are defined by the \code{\link{futile.logger}} package.
The following levels a available:
\tabular{ll}{
futile.logger::TRACE\cr 
futile.logger::DEBUG\cr
futile.logger::INFO\cr
futile.logger::WARN\cr
futile.logger::ERROR\cr
futile.logger::FATAL 
}
}
\keyword{internal}

