% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setMomentum<-}
\alias{setMomentum}
\alias{setMomentum<-}
\alias{setMomentum<-,Net-method}
\title{Sets the momentum of the \code{\link{Net}}.}
\usage{
setMomentum(net) <- value

\S4method{setMomentum}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{numeric}.}
}
\description{
Sets the momentum of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

