% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setNormalizeWeights<-,Net-method}
\alias{setNormalizeWeights<-,Net-method}
\title{Set whether weight normalization should be performed}
\usage{
\S4method{setNormalizeWeights}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Boolean value indicating whether the weights should be
normalized}
}
\description{
Set whether weight normalization should be performed
}
\seealso{
\link{setNormalizeWeights<-}
}

