% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\name{setNormalizeWeights<-}
\alias{setNormalizeWeights<-}
\title{Set whether weight normalization should be performed}
\usage{
setNormalizeWeights(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Boolean value indicating whether the weights should be
normalized}
}
\description{
Set whether weight normalization should be performed
}
\seealso{
\code{\link{Net}}
}

