\docType{methods}
\name{setNumHidden<-}
\alias{setNumHidden}
\alias{setNumHidden<-}
\alias{setNumHidden<-,RBM-method}
\title{Sets the number of hidden units}
\usage{
setNumHidden(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The number of hidden units}
}
\description{
Sets the number of hidden units
}
\seealso{
\code{\link{RBM}}
}

