\docType{methods}
\name{setNumVisible<-}
\alias{setNumVisible}
\alias{setNumVisible<-}
\alias{setNumVisible<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

