% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setOutput<-}
\alias{setOutput}
\alias{setOutput<-}
\alias{setOutput<-,RBM-method}
\title{Sets the output of the \code{\link{RBM}} object}
\usage{
setOutput(rbm) <- value

\S4method{setOutput}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The output of the \code{\link{RBM}} object}
}
\description{
Sets the output of the \code{\link{RBM}} object
}
\seealso{
\code{\link{RBM}}
}

