% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setRBMList<-}
\alias{setRBMList<-}
\alias{setRBMList<-,DArch-method}
\title{Sets the list of RBMs}
\usage{
setRBMList(darch) <- value

\S4method{setRBMList}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{The list of RBMs.}
}
\description{
Sets the list of RBMs
}
\seealso{
\code{\link{DArch}}
}

