\docType{methods}
\name{setVisibleBiases<-}
\alias{setVisibleBiases}
\alias{setVisibleBiases<-}
\alias{setVisibleBiases<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

