\docType{methods}
\name{setVisibleUnitStates<-}
\alias{setVisibleUnitStates}
\alias{setVisibleUnitStates<-}
\alias{setVisibleUnitStates<-,RBM-method}
\title{BEARBEITEN}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{BEARBEITEN}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

