\docType{methods}
\name{setWeightInc<-}
\alias{setWeightInc}
\alias{setWeightInc<-}
\alias{setWeightInc<-,RBM-method}
\title{Sets the update values for the weights}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The update values for the weights}
}
\description{
  Sets the update values for the weights
}
\seealso{
  \code{\link{RBM}}
}

