% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setWeights<-}
\alias{setWeights}
\alias{setWeights<-}
\alias{setWeights<-,RBM-method}
\title{Sets the weights of the \code{\link{RBM}} object}
\usage{
setWeights(rbm) <- value

\S4method{setWeights}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The weights of the \code{\link{RBM}} object.}
}
\description{
Sets the weights of the \code{\link{RBM}} object
}
\seealso{
\code{\link{RBM}}
}

