% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setWeightUpdateFunction<-}
\alias{setWeightUpdateFunction<-}
\alias{setWeightUpdateFunction<-,DArch-method}
\title{Set the weight update function for the given layer.}
\usage{
setDropoutMask(darch, i) <- value

\S4method{setWeightUpdateFunction}{DArch}(darch, layerIndex) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{layerIndex}{Layer index.}

\item{value}{Weight update function.}
}
\value{
The updated darch instance
}
\description{
Sets the function that is used to calculate the final weight updates (given
the updates from the fine-tune function).
}
\seealso{
\link{DArch}, \link{weightDecayWeightUpdate}
}

