% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show,DArch-method}
\alias{show,DArch-method}
\title{Print \code{\linkS4class{DArch}} details.}
\usage{
\S4method{show}{DArch}(object)
}
\arguments{
\item{object}{\code{\linkS4class{DArch}} instance}

\item{...}{Further parameters to \code{\link{print.DArch}}}
}
\description{
Simple redirection to \code{\link{print.DArch}}.
}
\keyword{internal}

