% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{tanhUnit}
\alias{tanhUnit}
\title{Continuous Tan-Sigmoid unit function.}
\usage{
tanhUnit(input, ...)
}
\arguments{
\item{input}{Input for the activation function.}
}
\value{
A list with the activation in the first entry and the derivative of
  the transfer function in the second entry
}
\description{
Calculates the unit activations and returns them in a list.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{linearUnit}},
  \code{\link{maxoutUnit}},
  \code{\link{rectifiedLinearUnit}},
  \code{\link{sigmoidUnit}}, \code{\link{softmaxUnit}},
  \code{\link{softplusUnit}}
}

