% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{tanSigmoidUnit}
\alias{tanSigmoidUnit}
\title{Continuous Tan-Sigmoid unit function.}
\usage{
tanSigmoidUnit(data, weights)
}
\arguments{
\item{data}{The data matrix for the calculation}

\item{weights}{The weight and bias matrix for the calculation}
}
\value{
A list with the activation of the transfer function in the first
 entry
}
\description{
Calculates the unit activations and returns them in a list.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{binSigmoidUnit}};
  \code{\link{linearUnitDerivative}};
  \code{\link{linearUnit}};
  \code{\link{maxoutUnitDerivative}};
  \code{\link{sigmoidUnitDerivative}};
  \code{\link{sigmoidUnit}};
  \code{\link{softmaxUnitDerivative}};
  \code{\link{softmaxUnit}};
  \code{\link{tanSigmoidUnitDerivative}}
}

