% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbmUnitFunctions.R
\name{tanSigmUnitFunc}
\alias{tanSigmUnitFunc}
\title{Calculates the neuron output with the hyperbolic tangent function}
\usage{
tanSigmUnitFunc(rbm, data, biases, weights, runParams,
  matMult = getDarchParam("matMult", `\%*\%`, ...), ...)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{biases}{The biases for the calculations}

\item{weights}{The weight matrix for the calculations}

\item{runParams}{Parameters which indicates the status of the training.}

\item{...}{Additional parameters}

\item{dataList}{A list with the data matrices for the calculations.}
}
\value{
The real value and binary (-1,1) activations for the units
}
\description{
Calculates the neuron output with the hyperbolic tangent function from
input in \code{data}.
}
\details{
The return value is a list with the output of the hyperbolic tangent
function as first entry and binary (-1,1) representation calculated through a
comparison ofthe output with random numbers. The random numbers a generated
with the function \code{\link{runif}}.
}
\seealso{
Other RBM unit functions: \code{\link{linearUnitFunc}};
  \code{\link{sigmUnitFunc}}
}

