\docType{methods}
\name{trainRBM}
\alias{trainRBM}
\alias{trainRBM,RBM-method}
\title{Train an RBM with Contrastiv Divergence}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{trainData}{The data for the training}

  \item{maxEpoch}{The number of training iterations}

  \item{numCD}{Number of contrastive divergence iterations}
}
\description{
  BEARBEITEN
}
\seealso{
  \code{\link{RBM}}
}

