% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{validateDataSet,DataSet-method}
\alias{validateDataSet,DataSet-method}
\title{Validate \code{\linkS4class{DataSet}}}
\usage{
\S4method{validateDataSet}{DataSet}(dataSet, darch)
}
\arguments{
\item{dataSet}{\code{\linkS4class{DataSet}} to validate}

\item{darch}{\code{\linkS4class{DArch}} object to validate this 
\code{\linkS4class{DataSet}} against.}
}
\description{
Validate \code{\linkS4class{DataSet}}
}
\seealso{
\link{validateDataSet}
}
\keyword{internal}

