% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{validateDataSet}
\alias{validateDataSet}
\title{Validate \code{\linkS4class{DataSet}}}
\usage{
validateDataSet(dataSet, darch)
}
\arguments{
\item{dataSet}{\code{\linkS4class{DataSet}} to validate}

\item{darch}{\code{\linkS4class{DArch}} object to validate this 
\code{\linkS4class{DataSet}} against.}
}
\value{
Logical indicating whether the \code{\linkS4class{DataSet}} is valid.
}
\description{
Validates the \code{\linkS4class{DataSet}} for the given
\code{\linkS4class{DArch}} object.
}
\details{
Validates the data dimensions and data types of the
\code{\linkS4class{DataSet}}.
}
\keyword{internal}

