% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{validateDataSets}
\alias{validateDataSets}
\title{Validates the \code{\link{DataSet}} for the given \code{\link{DArch}} object.}
\usage{
validateDataSets(dataSets, darch)
}
\arguments{
\item{darch}{\code{\link{DArch}} object to validate this
  \code{\link{DataSet}} against.}

\item{dataSet}{\code{\link{DataSet}} to validate}
}
\value{
Whether the \code{\link{DataSet}} is valid.
}
\description{
Validates the \code{\link{DataSet}} for the given \code{\link{DArch}} object.
}
\details{
Validates the data dimensions and data types of the
  \code{\link{DataSet}}.
}

