% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weightUpdateFunctions.R
\name{weightDecayWeightUpdate}
\alias{weightDecayWeightUpdate}
\title{Updates the weight using weight decay.}
\usage{
weightDecayWeightUpdate(darch, layerIndex, weightsInc, biasesInc)
}
\arguments{
\item{darch}{\linkS4class{DArch} instance.}

\item{layerIndex}{Layer index within the network.}

\item{weightsInc}{Matrix containing scheduled weight updates from the
fine-tuning algorithm.}

\item{biasesInc}{Bias weight updates.}

\item{weightDecay}{Weights are multiplied by (1 - \code{weightDecay}) before
each update. Corresponds to the \code{darch.weightDecay} parameter of
\link{darch.default}.}
}
\value{
updated \linkS4class{DArch} instance
}
\description{
Multiplies the weights by (1 - \code{weightDecay}) before applying the
scheduled weight changes.
}

