\docType{methods}
\name{addLayer}
\alias{addLayer}
\alias{addLayer,DArch-method}
\title{Adds a layer to the DArch object}
\arguments{
  \item{darch}{An instance of the class
  \code{\link{DArch}}.}
}
\description{
  Adds a layer to the given DArch object. The parameter
  weights and biases will be put together in one matrix.
}
\seealso{
  \code{\link{DArch}}, \code{\link{sigmoidUnit}},
  \code{\link{binSigmoidUnit}},
  \code{\link{sigmoidUnitDerivative}},
  \code{\link{linearUnit}},
  \code{\link{linearUnitDerivative}},
  \code{\link{softmaxUnit}},
  \code{\link{softmaxUnitDerivative}},
}

