\docType{methods}
\name{addLayerField}
\alias{addLayerField}
\alias{addLayerField,DArch-method}
\title{Adds a field to a layer}
\arguments{
  \item{darch}{An instance of the class
  \code{\link{DArch}}.}
}
\description{
  Adds a field to the layer given by the index.
}
\seealso{
  \code{\link{DArch}}
}

