% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{createDataSet,ANY,ANY,missing,missing-method}
\alias{createDataSet,ANY,ANY,missing,missing-method}
\alias{createDataSet.default}
\title{Create \code{\linkS4class{DataSet}} using data and targets.}
\usage{
\S4method{createDataSet}{ANY,ANY,missing,missing}(data, targets, ...,
  scale = F)
}
\arguments{
\item{data}{Input data, possibly also target data if a formula is used.}

\item{targets}{Target data.}

\item{...}{Further parameters.}

\item{scale}{Logical indicating whether to scale the data.}
}
\description{
Create \code{\linkS4class{DataSet}} using data and targets.
}
\seealso{
\link{createDataSet}
}

