% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{createDataSet,ANY,missing,formula,missing-method}
\alias{createDataSet,ANY,missing,formula,missing-method}
\alias{createDataSet.formula}
\title{Constructor function for \code{\linkS4class{DataSet}} objects.}
\usage{
\S4method{createDataSet}{ANY,missing,formula,missing}(data, formula, ...,
  na.action = na.pass, previous.dataSet = new("DataSet"))
}
\arguments{
\item{data}{\code{\link{data.frame}} containing the dataset.}

\item{formula}{Model formula.}

\item{...}{Additional parameters.}

\item{na.action}{\code{\link{model.frame}} parameter}

\item{previous.dataSet}{Existing \code{\linkS4class{DataSet}} from which
parameters are extrated and re-used.}
}
\value{
The new \code{\linkS4class{DataSet}} object
}
\description{
Generates a new \code{\linkS4class{DataSet}} object with the given 
parameters.
}
\details{
Initializes a \code{\linkS4class{DataSet}} with data using a
  \code{\link{formula}} and saving its parameters for subsequent data.
}
\seealso{
\link{darch.formula}, \link{createDataSet}
}
\keyword{internal}

