% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{createDataSet.DataSet}
\alias{createDataSet.DataSet}
\title{Create new \code{\linkS4class{DataSet}} by filling an existing one with new
data.}
\usage{
createDataSet.DataSet(data, targets, dataSet, ...)
}
\description{
Create new \code{\linkS4class{DataSet}} by filling an existing one with new
data.
}

