% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darch.R
\name{darch.formula}
\alias{darch.formula}
\title{Fit a deep neural network using a formula and a single data frame or matrix.}
\usage{
\method{darch}{formula}(x, data, layers, ..., dataValid = NULL,
  logLevel = NULL, paramsList = list())
}
\arguments{
\item{x}{The formula specifying the model.}

\item{data}{Data frame or matrix.}

\item{...}{additional parameters, see \link{darch.default}.}

\item{dataValid}{Data frame or matrix of validation data.}
}
\description{
Fit a deep neural network using a formula and a single data frame or matrix.
}
\seealso{
\link{model.frame}

Other darch interface functions: \code{\link{darch.DataSet}},
  \code{\link{darch.default}}, \code{\link{darchTest}},
  \code{\link{darch}}, \code{\link{predict.DArch}},
  \code{\link{print.DArch}}
}

