% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{darch.formula}
\alias{darch.formula}
\title{Fit a deep neural network using a formula and a single data frame or matrix.}
\usage{
\method{darch}{formula}(x, data, dataValid = NULL, ...)
}
\arguments{
\item{x}{The formula specifying the model.}

\item{data}{Data frame or matrix.}

\item{dataValid}{Data frame or matrix of validation data.}

\item{...}{additional parameters, see \link{darch.default}.}
}
\description{
Fit a deep neural network using a formula and a single data frame or matrix.
}
\seealso{
\link{model.frame}

Other darch interface functions: \code{\link{darch.DataSet}};
  \code{\link{darch.default}}; \code{\link{darch}};
  \code{\link{predict.DArch}}, \code{\link{predict.darch}};
  \code{\link{print.DArch}}, \code{\link{print.darch}}
}

