% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{darch}
\alias{darch}
\title{Fit a deep neural network.}
\usage{
darch(x, ...)
}
\arguments{
\item{x}{Input data.}

\item{...}{additional parameters}
}
\value{
Fitted \code{\linkS4class{DArch}} instance
}
\description{
Fit a deep neural network with optional pre-training and one of various
fine-tuning algorithms. See \link{darch.default} for a full list of
parameters.
}
\details{
The darch package implements Deep Architecture Networks and restricted
Boltzmann machines.

The creation of this package is motivated by the papers from G. Hinton et.
al. from 2006 (see references for details) and from the MATLAB source code
developed in this context. This package provides the possibility to generate
deep architecture networks (darch) like the deep belief networks from Hinton
et. al.. The deep architectures can then be trained with the contrastive
divergence method. After this pre-training it can be fine tuned with several
learning methods like backpropagation, resilient backpropagation and
conjugate gradients as well as more recent techniques like dropout and
maxout.

\tabular{ll}{ Package: \tab darch\cr Type: \tab Package\cr Version: \tab
0.9.2.9000\cr Date: \tab 2015-09-12\cr License: \tab GPL-2 or later\cr
LazyLoad: \tab yes\cr }
}
\examples{
source(paste0(system.file(package="darch"), "/examples/examples.R"))
}
\author{
Martin Drees \email{mdrees@stud.fh-dortmund.de} and contributors.
}
\references{
Hinton, G. E., S. Osindero, Y. W. Teh, A fast learning algorithm
  for deep belief nets, Neural Computation 18(7), S. 1527-1554, DOI:
  10.1162/neco.2006.18.7.1527 2006.

  Hinton, G. E., R. R. Salakhutdinov, Reducing the dimensionality of data
  with neural networks, Science 313(5786), S. 504-507, DOI:
  10.1126/science.1127647, 2006.

  Hinton, Geoffrey E. et al. (2012). "Improving neural networks by
  preventing coadaptation of feature detectors". In: Clinical Orthopaedics
  and Related Research abs/1207.0580. URL : http://arxiv.org/abs/1207.0580.

  Goodfellow, Ian J. et al. (2013). "Maxout Networks". In: Proceedings of
  the 30th International Conference on Machine Learning, ICML 2013, Atlanta,
  GA, USA, 16-21 June 2013, pp. 1319-1327.
  URL: http://jmlr.org/proceedings/papers/v28/goodfellow13.html.
}
\seealso{
Other darch interface functions: \code{\link{darch.DataSet}};
  \code{\link{darch.default}}; \code{\link{darch.formula}};
  \code{\link{predict.DArch}}, \code{\link{predict.darch}};
  \code{\link{print.DArch}}, \code{\link{print.darch}}
}
\keyword{Architectures}
\keyword{Backpropagation}
\keyword{Boltzmann}
\keyword{Conjugate}
\keyword{Contrastive}
\keyword{Deep}
\keyword{Deep-Belief-Networks}
\keyword{Divergence}
\keyword{Dropout}
\keyword{Gradient}
\keyword{Machines}
\keyword{Maxout}
\keyword{NN}
\keyword{Nets}
\keyword{Networks}
\keyword{Neural}
\keyword{Resilient}
\keyword{Restricted}
\keyword{darch}
\keyword{package}

