% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{darch}
\alias{darch}
\title{Fit a deep neural network.}
\usage{
darch(x, ...)
}
\description{
Fit a deep neural network with optional pre-training and one of various
fine-tuning algorithms.
}
\details{
The darch-package implements Deep Architecture Networks and restricted
Boltzmann machines.

The creation of this package is motivated by the papers from G. Hinton et.
al. from 2006 (see references for details) and from the MATLAB source code
developed in this context. This package provides the possibility to generate
deep architecture networks (darch) like the deep belief networks from Hinton
et. al.. The deep architectures can then be trained with the contrastive
divergence method. After this pre-training it can be fine tuned with several
learning methods like backpropagation, resilient backpropagation and
conjugate gradients. darch improves and extends darch, for example by
dropout and maxout.

\tabular{ll}{ Package: \tab darch\cr Type: \tab Package\cr Version: \tab
0.9.2.9000\cr Date: \tab 2015-07-14\cr License: \tab GPL-2 or later\cr
LazyLoad: \tab yes\cr }
}
\examples{
source(paste0(system.file(package="darch"), "/examples/examples.R"))
}
\author{
Martin Drees \email{mdrees@stud.fh-dortmund.de} and contributors.

  Maintainer: Johannes Rueckert \email{j.rueckert@gmx.net}
}
\references{
Hinton, G. E., S. Osindero, Y. W. Teh, A fast learning algorithm
  for deep belief nets, Neural Computation 18(7), S. 1527-1554, DOI:
  10.1162/neco.2006.18.7.1527 2006.

  Hinton, G. E., R. R. Salakhutdinov, Reducing the dimensionality of data
  with neural networks, Science 313(5786), S. 504-507, DOI:
  10.1126/science.1127647, 2006.
}
\keyword{Architectures}
\keyword{Backpropagation}
\keyword{Bolzmann}
\keyword{Conjugate}
\keyword{Contrastive}
\keyword{Deep}
\keyword{Deep-Belief-Networks}
\keyword{Divergence}
\keyword{Dropout}
\keyword{Gradient}
\keyword{Machines}
\keyword{Maxout}
\keyword{NN}
\keyword{Nets}
\keyword{Networks}
\keyword{Neural}
\keyword{Resilient}
\keyword{Restricted}
\keyword{TODO}
\keyword{add}
\keyword{darch}
\keyword{package}
\keyword{references}

