% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{darchModelInfo}
\alias{darchModelInfo}
\title{Creates a custom caret model for \code{darch}.}
\usage{
darchModelInfo(params = NULL, grid = NULL)
}
\arguments{
\item{params}{\code{\link{data.frame}} of parameters or \code{NULL} to use a
simple default (bp.learnRate).}

\item{grid}{Function which procuces a \code{\link{data.frame}} containing a
grid of parameter combinations or \code{NULL} to use a simple default.}
}
\value{
Valid \code{caret}  model which can be passed to
  \code{\link[caret]{train}}.
}
\description{
This function creates a \code{caret} model description to enable training
\code{DArch} instances with the \code{\link[caret]{train}} function. See the
documentation on custom caret models for further information and examples on
how to create valid \code{params} and \code{grid} values.
}
\seealso{
\href{https://topepo.github.io/caret/custom_models.html}{Caret custom models}
}

