\docType{methods}
\name{getBatchSize}
\alias{getBatchSize}
\alias{getBatchSize,Net-method}
\title{Returns the batch size of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns the batch size of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

