\docType{methods}
\name{getErrorFunction}
\alias{getErrorFunction}
\alias{getErrorFunction,Net-method}
\title{Returns the error function of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns the error function of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

