% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getErrorFunction}
\alias{getErrorFunction}
\alias{getErrorFunction,Net-method}
\title{Returns the error function of the \code{\link{Net}}.}
\usage{
getErrorFunction(net)

\S4method{getErrorFunction}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the error function of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

