\docType{methods}
\name{getExecuteFunction}
\alias{getExecuteFunction}
\alias{getExecuteFunction,DArch-method}
\title{Returns the function for the execution of the \code{\link{DArch}} network.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}
}
\description{
  Returns the function for the execution of the
  \code{\link{DArch}} network.
}
\seealso{
  \code{\link{DArch}}
}

