\docType{methods}
\name{getFF}
\alias{getFF}
\alias{getFF,Net-method}
\title{Returns if the weights are saved as ff objects}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns if the weights are saved as ff objects

  Returns the ff state of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}

  \code{\link{Net}}
}

