\docType{methods}
\name{getGenWeightFunction}
\alias{getGenWeightFunction}
\alias{getGenWeightFunction,Net-method}
\title{Returns the function for generating weight matrices.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns the function for generating weight matrices.
}
\seealso{
  \code{\link{Net}}
}

