% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getGenWeightFunction}
\alias{getGenWeightFunction}
\alias{getGenWeightFunction,Net-method}
\title{Returns the function for generating weight matrices.}
\usage{
getGenWeightFunction(net)

\S4method{getGenWeightFunction}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the function for generating weight matrices.
}
\seealso{
\code{\link{Net}}
}

