% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getHiddenBiases}
\alias{getHiddenBiases}
\alias{getHiddenBiases,RBM-method}
\title{Returns the biases of the hidden units.}
\usage{
getHiddenBiases(rbm)

\S4method{getHiddenBiases}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The biases of the hidden units.
}
\description{
Returns the biases of the hidden units.
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

