% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getHiddenUnitStates}
\alias{getHiddenUnitStates}
\alias{getHiddenUnitStates,RBM-method}
\title{Returns a list with the states of the hidden units.}
\usage{
getHiddenUnitStates(rbm)

\S4method{getHiddenUnitStates}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The states of the hidden units.
}
\description{
Returns a list with the states of the hidden units.
}
\seealso{
\code{\link{RBM}}
}

