% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getLearnRateWeights}
\alias{getLearnRateWeights}
\alias{getLearnRateWeights,Net-method}
\title{Returns the learn rate of the weights.}
\usage{
getLearnRateWeights(net)

\S4method{getLearnRateWeights}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
Returns the learn rate of the weights.
}
\seealso{
\code{\link{Net}}
}

