% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Getter.R
\docType{methods}
\name{getMomentum}
\alias{getMomentum}
\alias{getMomentum,Net-method}
\title{Returns the current momentum of the \code{\link{Net}}.}
\usage{
getMomentum(net)

\S4method{getMomentum}{Net}(net)
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{numEpochsRemaining}{The number of epochs the net is going to be
 trained for.}
}
\description{
Returns the current momentum of the \code{\link{Net}}.
}
\seealso{
\code{\link{Net}}
}

