% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.Getter.R
\name{getMomentum}
\alias{getMomentum}
\title{Returns the current momentum of the \code{Net}.}
\usage{
getMomentum(net)
}
\arguments{
\item{net}{A instance of the class \code{Net}.}
}
\description{
Returns the current momentum of the \code{Net}.
}
\seealso{
\code{\linkS4class{Net}}
}
\keyword{internal}

