\docType{methods}
\name{getMomentumSwitch}
\alias{getMomentumSwitch}
\alias{getMomentumSwitch,Net-method}
\title{Returns the momentum switch of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}
}
\description{
  Returns the momentum switch of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

