\docType{methods}
\name{getNumHidden}
\alias{getNumHidden}
\alias{getNumHidden,RBM-method}
\title{Returns the number of hidden units of the \code{\link{RBM}}}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The number of hidden units of the \code{\link{RBM}}
}
\description{
  Returns the number of hidden units of the
  \code{\link{RBM}}
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

