% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getNumVisible}
\alias{getNumVisible}
\alias{getNumVisible,RBM-method}
\title{Returns the number of visible units of the \code{\link{RBM}}}
\usage{
getNumVisible(rbm)

\S4method{getNumVisible}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The number of visible units of the \code{\link{RBM}}
}
\description{
Returns the number of visible units of the \code{\link{RBM}}
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

