\docType{methods}
\name{getOutput}
\alias{getOutput}
\alias{getOutput,RBM-method}
\title{Returns the output of the \code{\link{RBM}}}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The output of the \code{\link{RBM}}
}
\description{
  Returns the output of the \code{\link{RBM}}
}
\seealso{
  \code{\link{RBM}}
}

