\docType{methods}
\name{getPosPhaseData}
\alias{getPosPhaseData}
\alias{getPosPhaseData,RBM-method}
\title{Returns the data for the positive phaes.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The data for the positive phaes.
}
\description{
  Returns the data for the positive phaes.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

