\docType{methods}
\name{getVisibleBiases}
\alias{getVisibleBiases}
\alias{getVisibleBiases,RBM-method}
\title{Returns the biases of the visible units.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The biases of the visible units.
}
\description{
  Returns the biases of the visible units.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

