\docType{methods}
\name{getVisibleBiasesInc}
\alias{getVisibleBiasesInc}
\alias{getVisibleBiasesInc,RBM-method}
\title{Returns the update value for the biases of the visible units.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The update value for the biases of the visible units.
}
\description{
  Returns the update value for the biases of the visible
  units.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

