\docType{methods}
\name{getVisibleUnitStates}
\alias{getVisibleUnitStates}
\alias{getVisibleUnitStates,RBM-method}
\title{Returns a list with the states of the visible units.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The states of the visible units.
}
\description{
  Returns a list with the states of the visible units.
}
\seealso{
  \code{\link{RBM}}
}

