\docType{methods}
\name{getWeightCost}
\alias{getWeightCost}
\alias{getWeightCost,RBM-method}
\title{Returns the weigth cost for the training}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The weigth cost for the training
}
\description{
  Returns the weigth cost for the training
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

