% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getWeightCost}
\alias{getWeightCost}
\alias{getWeightCost,RBM-method}
\title{Returns the weigth cost for the training}
\usage{
getWeightCost(rbm)

\S4method{getWeightCost}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The weigth cost for the training
}
\description{
Returns the weigth cost for the training
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

