\docType{methods}
\name{getWeightInc}
\alias{getWeightInc}
\alias{getWeightInc,RBM-method}
\title{Returns the update value for the weights.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The update value for the weights.
}
\description{
  Returns the update value for the weights.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

