% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Getter.R
\docType{methods}
\name{getWeights}
\alias{getWeights}
\alias{getWeights,RBM-method}
\title{Returns the weights of the \code{\link{RBM}}.}
\usage{
getWeights(rbm)

\S4method{getWeights}{RBM}(rbm)
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
The weigths of the \code{\link{RBM}}.
}
\description{
Returns the weights of the \code{\link{RBM}}.
}
\seealso{
\code{\link{RBM}}

\code{\link{RBM}}
}

