\docType{methods}
\name{getWeights}
\alias{getWeights}
\alias{getWeights,RBM-method}
\title{Returns the weigths of the \code{\link{RBM}}.}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\value{
  The weigths of the \code{\link{RBM}}.
}
\description{
  Returns the weigths of the \code{\link{RBM}}.
}
\seealso{
  \code{\link{RBM}}

  \code{\link{RBM}}
}

