% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{linearUnitDerivative}
\alias{linearUnitDerivative}
\title{Linear unit function with unit derivatives.}
\usage{
linearUnitDerivative(input, ...)
}
\arguments{
\item{input}{Input for the activation function.}
}
\value{
A list with the linear activation in the first entry and the
derivative of the activation in the second entry
}
\description{
The function calculates the activation of the units and returns a list, in
which the first entry is the linear activation of the units and the second
entry is the derivative of the transfer function.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{maxoutUnitDerivative}};
  \code{\link{rectifiedLinearUnitDerivative}};
  \code{\link{sigmoidUnitDerivative}};
  \code{\link{softmaxUnitDerivative}};
  \code{\link{tanSigmoidUnitDerivative}}
}

