% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDArch.R
\docType{methods}
\name{loadDArch}
\alias{loadDArch}
\title{Loads a DArch network}
\usage{
loadDArch(name = "darch")
}
\arguments{
\item{name}{The name of the file without the ending ".net".}
}
\value{
\code{DArch} instance: the loaded deep architecture
}
\description{
Convenience function for loading \code{\linkS4class{DArch}} instances from
disk. There is no special re-initialization logic, so it is not necessary to
use this function, you can also just use \code{\link{load}}.
}
\details{
Loads the DArch object from the filename given through the parameter \code{name}
plus the ending ".net".

Make sure when you load a DArch object that every file written by 
the \code{\link{saveDArch}} function is in the working directory.
}
\seealso{
\code{\link{saveDArch}}
}
\keyword{internal}

