% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.R
\docType{class}
\name{Net-class}
\alias{Net}
\alias{Net-class}
\title{Abstract class for neural networks.}
\description{
This is a abstract class for neural networks. It provides some
functionalities used in more than one network type.
}
\section{Slots}{

\describe{
\item{\code{batchSize}}{Object of class \code{"numeric"}. The batch size
for the training and test data during the learning.}

\item{\code{errorFunction}}{Object of class \code{"function"}. Function for error
calculation.}

\item{\code{genWeightFunction}}{Object of class \code{"function"}. A function to
generate a randomly initialised weight matrix.}

\item{\code{normalizeWeights}}{Logical indicating whether to normalize the weights.}

\item{\code{normalizeWeightsBound}}{Upper bound on the L2 norm of incoming weight
vectors}

\item{\code{learnRate}}{Learning rate.}

\item{\code{learnRateScale}}{Scale for learning rate.}

\item{\code{weightDecay}}{Weight decay for the update of the weights. Weights will be
multiplied by (1 - \code{weightDecay}) during each weight update.}

\item{\code{initialMomentum}}{Initial momentum.}

\item{\code{finalMomentum}}{Final momentum.}

\item{\code{momentumRampLength}}{Defines the momentum ramp relative to the number of
epochs.}

\item{\code{epochs}}{Number of epochs.}

\item{\code{epochsScheduled}}{Number of epochs the \code{\link{Net}} will have been
trained for after the training, if the training is not aborted. Mainly to
keep track of training session lengths for resumed training.}

\item{\code{stats}}{Training statistics.}
}}
\author{
Martin Drees
}
\seealso{
\code{\linkS4class{DArch}}, \code{\linkS4class{RBM}}
}

