% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.Class.R
\docType{class}
\name{Net-class}
\alias{Net-class}
\title{Abstract class for neural networks.}
\description{
This is an abstract class for neural networks. It provides some 
functionalities used in more than one network type.
}
\section{Slots}{

\describe{
\item{\code{epochs}}{Number of epochs the network has been trained for.}

\item{\code{stats}}{Training statistics.}

\item{\code{parameters}}{List of parameters which do not change throughout training.}
}}
\author{
Martin Drees
}
\seealso{
Other darch classes: \code{\link{DArch-class}},
  \code{\link{DataSet-class}}, \code{\link{RBM-class}}
}
\keyword{internal}

