% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDArch.R
\name{newDArch}
\alias{newDArch}
\title{Constructor function for \code{\linkS4class{DArch}} objects.}
\usage{
newDArch(params)
}
\arguments{
\item{params}{Additional parameters as a list, also stored in
\code{darch@parameters}.}
}
\value{
The new DArch object
}
\description{
Generate a new \code{\linkS4class{DArch}} object with the given parameters.
}
\details{
This function is used internally only, please call \code{\link{darch}}
to create a new \code{\linkS4class{DArch}} instance.
}
\keyword{internal}

