% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newDArch.R
\name{newDArch}
\alias{newDArch}
\title{Constructor function for \code{\link{DArch}} objects.}
\usage{
newDArch(layers, batchSize, logLevel = INFO,
  genWeightFunc = generateWeightsRunif)
}
\arguments{
\item{layers}{Array of layer sizes.}

\item{batchSize}{Size of the batches}

\item{logLevel}{The logging level. See \code{\link{setLogLevel}} for details.}

\item{genWeightFunc}{The function for generating the weight matrices}
}
\value{
The new DArch object
}
\description{
Generate a new \code{\link{DArch}} object with the given parameters.
}
\details{
It is recommended to use this function for generating a new
\code{\link{DArch}} object, because this function generates and sets all the
necessary parameters like the internally used \code{\link{RBM}} networks,
the list of statistics (\code{stats}) etc.
}

