% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newRBM.R
\name{newRBM}
\alias{newRBM}
\title{Constructor function for RBM object.}
\usage{
newRBM(numVisible, numHidden, batchSize, logLevel = INFO,
  genWeightFunc = generateWeightsRunif)
}
\arguments{
\item{numVisible}{Number of visible units.}

\item{numHidden}{Number of hidden units.}

\item{batchSize}{Size of the batches}

\item{logLevel}{The logging level.
See \code{\link{setLogLevel}} for details.}

\item{genWeightFunc}{The function for generating the weight matrices}
}
\value{
The new RBM object
}
\description{
TODO: Doc ...
}

