% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{predict.DArch}
\alias{predict.DArch}
\alias{predict.darch}
\title{Forward-propagate data.}
\usage{
\method{predict}{DArch}(darch, newdata = NULL, type = "raw")
}
\arguments{
\item{darch}{\code{\linkS4class{DArch}} instance}

\item{newdata}{New data to predict,  to return latest network output}

\item{type}{Output type, one of: \code{raw}, \code{bin}, \code{class}.}
}
\value{
Vector or matrix of networks outputs, output type depending on the
  \code{type} parameter
}
\description{
Forward-propagate given data through the deep neural network.
}

