% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.DArch}
\alias{print.DArch}
\alias{print.darch}
\title{Print \code{\linkS4class{DArch}} details.}
\usage{
\method{print}{DArch}(x, ...)
}
\arguments{
\item{x}{\code{\linkS4class{DArch}} instance}

\item{...}{Further parameters, not used.}
}
\description{
Print verbose information about a \linkS4class{DArch} instance.
}
\seealso{
Other darch interface functions: \code{\link{darch.DataSet}},
  \code{\link{darch.default}}, \code{\link{darch.formula}},
  \code{\link{darch}}, \code{\link{predict.DArch}},
  \code{\link{testDarch}}
}

