% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mnist.R
\name{provideMNIST}
\alias{provideMNIST}
\title{Provides MNIST data set in the given folder.}
\usage{
provideMNIST(folder = "data/")
}
\arguments{
\item{folder}{Folder name, including a trailing slash.}
}
\value{
Boolean value indicating success or failure.
}
\description{
This function will, if necessary, download the compressed MNIST data set and
convert it to \code{ff} files using \code{\link{readMNIST}}.
}

