% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mnist.R
\name{provideMNIST}
\alias{provideMNIST}
\title{Provides MNIST data set in the given folder.}
\usage{
provideMNIST(folder = "data/", download = F)
}
\arguments{
\item{folder}{Folder name, including a trailing slash.}

\item{download}{Logical indicating whether download is allowed.}
}
\value{
Boolean value indicating success or failure.
}
\description{
This function will, if necessary and allowed, download the compressed MNIST
data set and save it to .RData files using \code{\link{readMNIST}}.
}

