% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorFunctions.R
\name{quadraticError}
\alias{quadraticError}
\title{Quadratic error function}
\usage{
quadraticError(original, estimate)
}
\arguments{
\item{original}{The original data matrix}

\item{estimate}{The calculated data matrix}
}
\value{
A list with the name of the error function in the first entry and the
error value in the second entry
}
\description{
The function calculates the quadratic error from the \code{original} and
\code{estimate} parameters.
}
\seealso{
Other error functions: \code{\link{crossEntropyError}};
  \code{\link{mseError}}
}

