% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.R
\docType{class}
\name{RBM-class}
\alias{RBM}
\alias{RBM-class}
\title{Class for restricted Boltzmann machines}
\description{
This class represents a restricted Boltzmann machine.
}
\details{
For generating a RBM it is recommended to use the constructor function:
\code{\link{newRBM}}. The RBM can be trained with the implementation of the
contrastive divergence method \code{\link{trainRBM}}. The class inherits the
attributes from the \code{\linkS4class{Net}}.
}
\section{Slots}{

\describe{
\item{\code{numHidden}}{Object of class \code{"numeric"}. Number of hidden units.}

\item{\code{numVisible}}{Object of class \code{"numeric"}. Number of visible units.}

\item{\code{weights}}{Object of class \code{"matrix"}. Weight matrix.}

\item{\code{weightInc}}{Object of class \code{"matrix"}. Matrix of update values for
the Weight.}

\item{\code{output}}{Object of class \code{"matrix"}. Output matrix of the RBM.}

\item{\code{visibleBiases}}{Object of class \code{"array"}. Visible biases array.}

\item{\code{visibleBiasesInc}}{Object of class \code{"array"}. Array of update values
for the visible biases}

\item{\code{visibleUnitFunction}}{Object of class \code{"function"}. Unit function
for the visible units.}

\item{\code{visibleUnitStates}}{Object of class \code{"list"}. States of the visible
units.}

\item{\code{hiddenBiases}}{Object of class \code{"array"}. Hidden biases array.}

\item{\code{hiddenBiasesInc}}{Object of class \code{"array"}. Array of update values
for the hidden biases.}

\item{\code{hiddenUnitFunction}}{Object of class \code{"function"}. Unit function for
the hidden units.}

\item{\code{hiddenUnitStates}}{Object of class \code{"list"}. States of the hidden
units.}

\item{\code{updateFunction}}{Object of class \code{"function"}. Function for updating
the weights and biases.}

\item{\code{posPhaseData}}{Object of class \code{"list"}. Attribute to save the
positive phase data during the training.}
}}
\author{
Martin Drees
}
\seealso{
\code{\linkS4class{Net}}, \code{\linkS4class{DArch}},
  \code{\link{trainRBM}}
}

