\name{RBM}
\alias{RBM}
\alias{RBM-class}
\title{Class for Restricted-Bolzmann-Machine}
\description{
  To the constructor function: \code{\link{newRBM}} TODO:
  Description ...
}
\details{
  TODO: Details ...
}
\section{Slots}{
  \describe{ \item{\code{rbmList}:}{Object of class
  \code{"list"}. Contains all rbm's for the pre-training.}
  }
}
\author{
  Martin Drees
}
\seealso{
  \code{\linkS4class{Net}}, \code{\linkS4class{DArch}}
}

