% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbmUpdate.R
\name{rbmUpdate}
\alias{rbmUpdate}
\title{Function for updating the weights and biases of an \code{RBM}}
\usage{
rbmUpdate(rbm, matMult = getParameter(".matMult", net = rbm))
}
\arguments{
\item{rbm}{A instance of the class \code{\linkS4class{RBM}}.}

\item{matMult}{Matrix multiplication function.}

\item{...}{Additional arguments.}
}
\value{
The updated \code{\linkS4class{RBM}}.
}
\description{
This function updates the weights and biases for an \code{\linkS4class{RBM}} 
network. It is saved in the attribute \code{updateFunction} of the 
\code{RBM} object and called from the training function 
\code{\link{trainRBM}}.
}
\seealso{
\code{\linkS4class{RBM}}
}
\keyword{internal}

