% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mnist.R
\name{readMNIST}
\alias{readMNIST}
\title{Function for generating .RData files of the MNIST Database}
\usage{
readMNIST(folder)
}
\arguments{
\item{folder}{The location of the MNIST-Database files.}
}
\description{
This function reads the MNIST-Database, randomized it and saves it in the
files "train" for the training data and "test" for test data.
}
\details{
When the data is read the variables for the training data is \code{trainData}
and \code{trainLabels} and for the test data \code{testData} and
\code{testLabels}. To start the function
The files "train-images-idx3-ubyte", "train-labels-idx1-ubyte',
"t10k-images-idx3-ubyte", and "t10k-labels-idx1-ubyte" have to be in the
folder given by the parameter \code{folder}. The folder name must end with
a slash.
}

