% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Remove.R
\docType{methods}
\name{removeLayerField}
\alias{removeLayerField}
\alias{removeLayerField,DArch-method}
\title{Removes a layer from the \code{\link{DArch}} object}
\usage{
removeLayerField(darch, index)

\S4method{removeLayerField}{DArch}(darch, index)
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{index}{The index of the layer.}
}
\value{
The \code{\link{DArch}} object without the layer.
}
\description{
This function removes the layer with the given index from the
\code{\link{DArch}} object.
}
\seealso{
\code{\link{DArch}}
}

