% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Reset.R
\docType{methods}
\name{resetExecOutput}
\alias{resetExecOutput}
\alias{resetExecOutput,DArch-method}
\title{Resets the output list of the \code{\link{DArch}} object}
\usage{
resetExecOutput(darch)

\S4method{resetExecOutput}{DArch}(darch)
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}
}
\description{
This function sets the attribute \code{executeOutput} of the
\code{\link{DArch}} object to an empty list.
}
\seealso{
\code{\link{DArch}}
}

