\docType{methods}
\name{resetRBM}
\alias{resetRBM}
\alias{resetRBM,RBM-method}
\title{Resets the weights and biases of the \code{\link{RBM}} object}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}
}
\description{
  This function resets the weights and biases of the
  \code{\link{RBM}} object.
}
\seealso{
  \code{\link{RBM}}
}

