% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runDArch.R
\name{runDArchDropout}
\alias{runDArchDropout}
\title{Execute the darch with dropout support}
\usage{
runDArchDropout(darch, data,
  iterations = getDarchParam("darch.dropout.momentMatching", 0, darch =
  darch), matMult = getDarchParam("matMult", `\%*\%`, darch = darch))
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{data}{The input data to execute the darch on.}
}
\value{
The DArch object with the calculated outputs
}
\description{
If dropout was disabled, \code{\link{runDArch}} will be called instead.
}
\seealso{
\code{\link{DArch}}

Other darch execute functions: \code{\link{runDArch}}
}

