% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDArch.R
\name{saveDArch}
\alias{saveDArch}
\title{Saves a DArch network}
\usage{
saveDArch(darch, name = "darch", trim = F)
}
\arguments{
\item{darch}{An instance of the class \code{\linkS4class{DArch}}.}

\item{name}{The name for the file. Default value is "darch".}

\item{trim}{Whether to trim the model before saving.}
}
\description{
Saves the DArch object to the filename given through the parameter \code{name}
plus the ending ".net".
}
\details{
If the parameter \code{saveRBM} is \code{FALSE} the field
  \code{rbmList} of the DArch object is overwritten by an empty list.
}
\seealso{
\code{\link{loadDArch}}
}
\keyword{internal}

