% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveRBM.R
\docType{methods}
\name{saveRBM}
\alias{saveRBM}
\alias{saveRBM,RBM-method}
\title{Saves a RBM network}
\usage{
saveRBM(rbm,name="rbm")

\S4method{saveRBM}{RBM}(rbm, name = "rbm")
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{name}{The name for the file. Default value is "rbm".}
}
\description{
Saves the RBM object to the filename given through the parameter \code{name}
plus the ending ".net".
}
\seealso{
\code{\link{loadRBM}}
}

