% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setCancel<-}
\alias{setCancel}
\alias{setCancel<-}
\alias{setCancel<-,DArch-method}
\title{Set whether the learning shall be canceled.}
\usage{
setCancel(darch) <- value

\S4method{setCancel}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{Boolean value if the learning shall canceled}
}
\description{
Set whether the learning shall be canceled.
}
\seealso{
\code{\link{DArch}}
}

