\docType{methods}
\name{setCancelMessage<-}
\alias{setCancelMessage}
\alias{setCancelMessage<-}
\alias{setCancelMessage<-,DArch-method}
\title{Sets the cancel message.}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{The message for the termination}
}
\description{
  Sets the cancel message.
}
\seealso{
  \code{\link{DArch}}
}

