% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setExecuteFunction<-}
\alias{setExecuteFunction}
\alias{setExecuteFunction<-}
\alias{setExecuteFunction<-,DArch-method}
\title{Sets the execution function for the network}
\usage{
setExecuteFunction(darch)  <- value

\S4method{setExecuteFunction}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{The execution function for the network.}
}
\description{
Sets the execution function for the network
}
\seealso{
\code{\link{DArch}}
}

