\docType{methods}
\name{setExecuteFunction<-}
\alias{setExecuteFunction}
\alias{setExecuteFunction<-}
\alias{setExecuteFunction<-,DArch-method}
\title{Sets the execution function for the network}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{The execution function for the network.}
}
\description{
  Sets the execution function for the network
}
\seealso{
  \code{\link{DArch}}
}

