\docType{methods}
\name{setFF<-}
\alias{setFF}
\alias{setFF<-}
\alias{setFF<-,Net-method}
\title{Sets if the weights are saved as ff objects}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Boolean value which indicates if the weights
  are saved as ff objects}
}
\description{
  Sets if the weights are saved as ff objects
}
\seealso{
  \code{\link{Net}}
}

