\docType{methods}
\name{setFinalMomentum<-}
\alias{setFinalMomentum}
\alias{setFinalMomentum<-}
\alias{setFinalMomentum<-,Net-method}
\title{Sets the final momentum of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class \code{numeric}.}
}
\description{
  Sets the final momentum of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

