% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setFineTuneFunction<-}
\alias{setFineTuneFunction}
\alias{setFineTuneFunction<-}
\alias{setFineTuneFunction<-,DArch-method}
\title{Sets the fine tuning function for the network}
\usage{
setFineTuneFunction(darch)  <- value

\S4method{setFineTuneFunction}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{The fine tuning function for the network.}
}
\description{
Sets the fine tuning function for the network
}
\seealso{
\code{\link{DArch}}
}

