% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.Setter.R
\docType{methods}
\name{setGenWeightFunction<-}
\alias{setGenWeightFunction}
\alias{setGenWeightFunction<-}
\alias{setGenWeightFunction<-,Net-method}
\title{Sets the function for generating weight matrices.}
\usage{
setGenWeightFunction(net) <- value

\S4method{setGenWeightFunction}{Net}(net) <- value
}
\arguments{
\item{net}{A instance of the class \code{\link{Net}}.}

\item{value}{Object of the class \code{function}.}
}
\description{
The function have to return a matrix with number of units in the lower layer
as number of rows and number of units in the upper layer as the number of
columns.
}
\seealso{
\code{\link{Net}}
}

