\docType{methods}
\name{setHiddenBiases<-}
\alias{setHiddenBiases}
\alias{setHiddenBiases<-}
\alias{setHiddenBiases<-,RBM-method}
\title{Sets the biases of the hidden units for the \code{\link{RBM}} object}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The biases of the hidden units for the
  \code{\link{RBM}} object.}
}
\description{
  Sets the biases of the hidden units for the
  \code{\link{RBM}} object
}
\seealso{
  \code{\link{RBM}}
}

