% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setHiddenBiases<-}
\alias{setHiddenBiases}
\alias{setHiddenBiases<-}
\alias{setHiddenBiases<-,RBM-method}
\title{Sets the biases of the hidden units for the \code{\link{RBM}} object}
\usage{
setHiddenBiases(rbm) <- value

\S4method{setHiddenBiases}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The biases of the hidden units for the \code{\link{RBM}} object.}
}
\description{
Sets the biases of the hidden units for the \code{\link{RBM}} object
}
\seealso{
\code{\link{RBM}}
}

