\docType{methods}
\name{setHiddenBiasesInc<-}
\alias{setHiddenBiasesInc}
\alias{setHiddenBiasesInc<-}
\alias{setHiddenBiasesInc<-,RBM-method}
\title{Sets the update value for the biases of the hidden units}
\usage{
setHiddenBiasesInc(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The update value for the biases of the
  hidden units.}
}
\description{
Sets the update value for the biases of the hidden units
}
\seealso{
\code{\link{RBM}}
}

