% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setHiddenBiasesInc<-}
\alias{setHiddenBiasesInc}
\alias{setHiddenBiasesInc<-}
\alias{setHiddenBiasesInc<-,RBM-method}
\title{Sets the update value for the biases of the hidden units}
\usage{
setHiddenBiasesInc(rbm) <- value

\S4method{setHiddenBiasesInc}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The update value for the biases of the hidden units.}
}
\description{
Sets the update value for the biases of the hidden units
}
\seealso{
\code{\link{RBM}}
}

