% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setHiddenUnitFunction<-}
\alias{setHiddenUnitFunction}
\alias{setHiddenUnitFunction<-}
\alias{setHiddenUnitFunction<-,RBM-method}
\title{Sets the unit function of the hidden units}
\usage{
setHiddenUnitFunction(rbm) <- value

\S4method{setHiddenUnitFunction}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The unit function of the hidden units}
}
\description{
Sets the unit function of the hidden units
}
\seealso{
\code{\link{RBM}}
}

