% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setHiddenUnitStates<-}
\alias{setHiddenUnitStates}
\alias{setHiddenUnitStates<-}
\alias{setHiddenUnitStates<-,RBM-method}
\title{Sets the states of the hidden units}
\usage{
setHiddenUnitStates(rbm) <- value

\S4method{setHiddenUnitStates}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The states of the hidden units}
}
\description{
Sets the states of the hidden units
}
\seealso{
\code{\link{RBM}}
}

