\docType{methods}
\name{setHiddenUnitStates<-}
\alias{setHiddenUnitStates}
\alias{setHiddenUnitStates<-}
\alias{setHiddenUnitStates<-,RBM-method}
\title{Sets the states of the hidden units}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The states of the hidden units}
}
\description{
  Sets the states of the hidden units
}
\seealso{
  \code{\link{RBM}}
}

