\docType{methods}
\name{setLayerFunction<-}
\alias{setLayerFunction}
\alias{setLayerFunction<-}
\alias{setLayerFunction<-,DArch-method}
\title{Sets the function for a layer with the given index}
\usage{
setLayerFunction(darch,index) <- value
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{index}{The index of the layer.}

  \item{value}{The function for the layer.}
}
\description{
Sets the function for a layer with the given index
}
\seealso{
\code{\link{DArch}}
}

