% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setLayers<-}
\alias{setLayers}
\alias{setLayers<-}
\alias{setLayers<-,DArch-method}
\title{Sets the layers for the network}
\usage{
setLayers(darch) <- value

\S4method{setLayers}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{The layers for the network.}
}
\description{
Sets the layers for the network
}
\seealso{
\code{\link{DArch}}
}

