\docType{methods}
\name{setLayerWeights<-}
\alias{setLayerWeights}
\alias{setLayerWeights<-}
\alias{setLayerWeights<-,DArch-method}
\title{Sets the weights of a layer with the given index}
\usage{
setLayerWeights(darch,index) <- value
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{index}{The index of the layer.}

  \item{value}{The weights for the layer.}
}
\description{
Sets the weights of a layer with the given index
}
\seealso{
\code{\link{DArch}}
}

