% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setLayerWeights<-}
\alias{setLayerWeights}
\alias{setLayerWeights<-}
\alias{setLayerWeights<-,DArch-method}
\title{Sets the weights of a layer with the given index}
\usage{
setLayerWeights(darch,index) <- value

\S4method{setLayerWeights}{DArch}(darch, index) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{index}{The index of the layer.}

\item{value}{The weights for the layer.}
}
\description{
Sets the weights of a layer with the given index
}
\seealso{
\code{\link{DArch}}
}

