\docType{methods}
\name{setLearnRateBiases<-}
\alias{setLearnRateBiases}
\alias{setLearnRateBiases<-}
\alias{setLearnRateBiases<-,DArch-method}
\title{Sets the learning rate for the biases}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{The learning rate for the biases.}
}
\description{
  Sets the learning rate for the biases
}
\seealso{
  \code{\link{DArch}}
}

