% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darch.Setter.R
\docType{methods}
\name{setLearnRateBiases<-}
\alias{setLearnRateBiases}
\alias{setLearnRateBiases<-}
\alias{setLearnRateBiases<-,DArch-method}
\title{Sets the learning rate for the biases}
\usage{
setLearnRateBiases(darch)  <- value

\S4method{setLearnRateBiases}{DArch}(darch) <- value
}
\arguments{
\item{darch}{A instance of the class \code{\link{DArch}}.}

\item{value}{The learning rate for the biases.}
}
\description{
Sets the learning rate for the biases
}
\seealso{
\code{\link{DArch}}
}

