\docType{methods}
\name{setLearnRateBiasHidden<-}
\alias{setLearnRateBiasHidden}
\alias{setLearnRateBiasHidden<-}
\alias{setLearnRateBiasHidden<-,RBM-method}
\title{Sets the learning rates of the biases for the hidden units}
\usage{
setLearnRateBiasHidden(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The learning rates of the biases for the
  hidden units}
}
\description{
Sets the learning rates of the biases for the hidden units
}
\seealso{
\code{\link{RBM}}
}

