\docType{methods}
\name{setLearnRateBiasHidden<-}
\alias{setLearnRateBiasHidden}
\alias{setLearnRateBiasHidden<-}
\alias{setLearnRateBiasHidden<-,RBM-method}
\title{Sets the learnig rates of the biases for the hidden units}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The learnig rates of the biases for the
  hidden units}
}
\description{
  Sets the learnig rates of the biases for the hidden units
}
\seealso{
  \code{\link{RBM}}
}

