\docType{methods}
\name{setLearnRateBiasVisible<-}
\alias{setLearnRateBiasVisible}
\alias{setLearnRateBiasVisible<-}
\alias{setLearnRateBiasVisible<-,RBM-method}
\title{Sets the learnig rates of the biases for the  visible units}
\usage{
setLearnRateBiasVisible(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The learnig rates of the biases for the
  visible units}
}
\description{
Sets the learnig rates of the biases for the visible units
}
\seealso{
\code{\link{RBM}}
}

