% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setLearnRateBiasVisible<-}
\alias{setLearnRateBiasVisible}
\alias{setLearnRateBiasVisible<-}
\alias{setLearnRateBiasVisible<-,RBM-method}
\title{Sets the learnig rates of the biases for the  visible units}
\usage{
setLearnRateBiasVisible(rbm) <- value

\S4method{setLearnRateBiasVisible}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The learnig rates of the biases for the visible units}
}
\description{
Sets the learnig rates of the biases for the  visible units
}
\seealso{
\code{\link{RBM}}
}

