\docType{methods}
\name{setLearnRateWeights<-}
\alias{setLearnRateWeights}
\alias{setLearnRateWeights<-}
\alias{setLearnRateWeights<-,Net-method}
\title{Sets the learning rate for the weights.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class \code{numeric}.}
}
\description{
  Sets the learning rate for the weights.
}
\seealso{
  \code{\link{Net}}
}

