\docType{methods}
\name{setLogLevel<-}
\alias{setLogLevel}
\alias{setLogLevel<-}
\alias{setLogLevel<-,Net-method}
\title{Sets the log level for the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class \code{numeric}.}
}
\description{
  The log levels a defined by the \code{\link{log4r}}
  package. The following levels a available: \tabular{ll}{
  log4r:::DEBUG = 1\cr log4r:::INFO = 2\cr log4r:::WARN =
  3\cr log4r:::ERROR = 4\cr log4r:::FATAL = 5 }
}
\seealso{
  \code{\link{Net}}
}

