\docType{methods}
\name{setMomentum<-}
\alias{setMomentum}
\alias{setMomentum<-}
\alias{setMomentum<-,Net-method}
\title{Sets the momentum of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class \code{numeric}.}
}
\description{
  Sets the momentum of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

