\docType{methods}
\name{setMomentumSwitch<-}
\alias{setMomentumSwitch}
\alias{setMomentumSwitch<-}
\alias{setMomentumSwitch<-,Net-method}
\title{Sets the momentum switch of the \code{\link{Net}}.}
\arguments{
  \item{net}{A instance of the class \code{\link{Net}}.}

  \item{value}{Object of the class \code{numeric}.}
}
\description{
  Sets the momentum switch of the \code{\link{Net}}.
}
\seealso{
  \code{\link{Net}}
}

