% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setNumHidden<-}
\alias{setNumHidden}
\alias{setNumHidden<-}
\alias{setNumHidden<-,RBM-method}
\title{Sets the number of hidden units}
\usage{
setNumHidden(rbm) <- value

\S4method{setNumHidden}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The number of hidden units}
}
\description{
Sets the number of hidden units
}
\seealso{
\code{\link{RBM}}
}

