% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setNumVisible<-}
\alias{setNumVisible}
\alias{setNumVisible<-}
\alias{setNumVisible<-,RBM-method}
\title{Sets the number of visible units}
\usage{
setNumVisible(rbm) <- value

\S4method{setNumVisible}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The number of visible units}
}
\description{
Sets the number of visible units
}
\seealso{
\code{\link{RBM}}
}

