\docType{methods}
\name{setNumVisible<-}
\alias{setNumVisible}
\alias{setNumVisible<-}
\alias{setNumVisible<-,RBM-method}
\title{Sets the number of visible units}
\usage{
setNumVisible(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The number of visible units}
}
\description{
Sets the number of visible units
}
\seealso{
\code{\link{RBM}}
}

