\docType{methods}
\name{setOutput<-}
\alias{setOutput}
\alias{setOutput<-}
\alias{setOutput<-,RBM-method}
\title{Sets the output of the \code{\link{RBM}} object}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The output of the \code{\link{RBM}} object}
}
\description{
  Sets the output of the \code{\link{RBM}} object
}
\seealso{
  \code{\link{RBM}}
}

