\docType{methods}
\name{setPosPhaseData<-}
\alias{setPosPhaseData}
\alias{setPosPhaseData<-}
\alias{setPosPhaseData<-,RBM-method}
\title{Sets the positive phase data for the training}
\usage{
setPosPhaseData(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The positive phase data for the training}
}
\description{
Sets the positive phase data for the training
}
\seealso{
\code{\link{RBM}}
}

