% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setPosPhaseData<-}
\alias{setPosPhaseData}
\alias{setPosPhaseData<-}
\alias{setPosPhaseData<-,RBM-method}
\title{Sets the positive phase data for the training}
\usage{
setPosPhaseData(rbm) <- value

\S4method{setPosPhaseData}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The positive phase data for the training}
}
\description{
Sets the positive phase data for the training
}
\seealso{
\code{\link{RBM}}
}

