\docType{methods}
\name{setRBMList<-}
\alias{setRBMList}
\alias{setRBMList<-}
\alias{setRBMList<-,DArch-method}
\title{Sets the list of RBMs}
\usage{
setRBMList(darch) <- value
}
\arguments{
  \item{darch}{A instance of the class
  \code{\link{DArch}}.}

  \item{value}{The list of RBMs.}
}
\description{
Sets the list of RBMs
}
\seealso{
\code{\link{DArch}}
}

