% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setUpdateFunction<-}
\alias{setUpdateFunction}
\alias{setUpdateFunction<-}
\alias{setUpdateFunction<-,RBM-method}
\title{Sets the update function of the \code{\link{RBM}} object}
\usage{
setUpdateFunction(rbm) <- value

\S4method{setUpdateFunction}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The update function of the \code{\link{RBM}} object.}
}
\description{
Sets the update function of the \code{\link{RBM}} object
}
\seealso{
\code{\link{RBM}}
}

