\docType{methods}
\name{setVisibleBiases<-}
\alias{setVisibleBiases}
\alias{setVisibleBiases<-}
\alias{setVisibleBiases<-,RBM-method}
\title{Sets the biases of the visible units for the \code{\link{RBM}} object}
\usage{
setVisibleBiases(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The biases of the visible units for the
  \code{\link{RBM}} object.}
}
\description{
Sets the biases of the visible units for the
\code{\link{RBM}} object
}
\seealso{
\code{\link{RBM}}
}

