\docType{methods}
\name{setVisibleBiasesInc<-}
\alias{setVisibleBiasesInc}
\alias{setVisibleBiasesInc<-}
\alias{setVisibleBiasesInc<-,RBM-method}
\title{Sets the update value for the biases of the visible units}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The update value for the biases of the
  visible units.}
}
\description{
  Sets the update value for the biases of the visible units
}
\seealso{
  \code{\link{RBM}}
}

