% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setVisibleBiasesInc<-}
\alias{setVisibleBiasesInc}
\alias{setVisibleBiasesInc<-}
\alias{setVisibleBiasesInc<-,RBM-method}
\title{Sets the update value for the biases of the visible units}
\usage{
setVisibleBiasesInc(rbm) <- value

\S4method{setVisibleBiasesInc}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The update value for the biases of the visible units.}
}
\description{
Sets the update value for the biases of the visible units
}
\seealso{
\code{\link{RBM}}
}

