% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setVisibleUnitFunction<-}
\alias{setVisibleUnitFunction}
\alias{setVisibleUnitFunction<-}
\alias{setVisibleUnitFunction<-,RBM-method}
\title{Sets the unit function of the visible units}
\usage{
setVisibleUnitFunction(rbm) <- value

\S4method{setVisibleUnitFunction}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The unit function of the visible units}
}
\description{
Sets the unit function of the visible units
}
\seealso{
\code{\link{RBM}}
}

