\docType{methods}
\name{setVisibleUnitFunction<-}
\alias{setVisibleUnitFunction}
\alias{setVisibleUnitFunction<-}
\alias{setVisibleUnitFunction<-,RBM-method}
\title{Sets the unit function of the visible units}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The unit function of the visible units}
}
\description{
  Sets the unit function of the visible units
}
\seealso{
  \code{\link{RBM}}
}

