% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setVisibleUnitStates<-}
\alias{setVisibleUnitStates}
\alias{setVisibleUnitStates<-}
\alias{setVisibleUnitStates<-,RBM-method}
\title{Sets the states of the visible units}
\usage{
setVisibleUnitStates(rbm) <- value

\S4method{setVisibleUnitStates}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The states of the visible units}
}
\description{
Sets the states of the visible units
}
\seealso{
\code{\link{RBM}}
}

