\docType{methods}
\name{setVisibleUnitStates<-}
\alias{setVisibleUnitStates}
\alias{setVisibleUnitStates<-}
\alias{setVisibleUnitStates<-,RBM-method}
\title{Sets the states of the visible units}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The states of the visible units}
}
\description{
  Sets the states of the visible units
}
\seealso{
  \code{\link{RBM}}
}

