% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setWeightCost<-}
\alias{setWeightCost}
\alias{setWeightCost<-}
\alias{setWeightCost<-,RBM-method}
\title{Sets the weight costs for the training}
\usage{
setWeightCost(rbm) <- value

\S4method{setWeightCost}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The weight costs for the training}
}
\description{
Sets the weight costs for the training
}
\seealso{
\code{\link{RBM}}
}

