\docType{methods}
\name{setWeightCost<-}
\alias{setWeightCost}
\alias{setWeightCost<-}
\alias{setWeightCost<-,RBM-method}
\title{Sets the weight costs for the training}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The weight costs for the training}
}
\description{
  Sets the weight costs for the training
}
\seealso{
  \code{\link{RBM}}
}

