% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.Setter.R
\docType{methods}
\name{setWeightInc<-}
\alias{setWeightInc}
\alias{setWeightInc<-}
\alias{setWeightInc<-,RBM-method}
\title{Sets the update values for the weights}
\usage{
setWeightInc(rbm) <- value

\S4method{setWeightInc}{RBM}(rbm) <- value
}
\arguments{
\item{rbm}{A instance of the class \code{\link{RBM}}.}

\item{value}{The update values for the weights}
}
\description{
Sets the update values for the weights
}
\seealso{
\code{\link{RBM}}
}

