\docType{methods}
\name{setWeights<-}
\alias{setWeights}
\alias{setWeights<-}
\alias{setWeights<-,RBM-method}
\title{Sets the weights of the \code{\link{RBM}} object}
\usage{
setWeights(rbm) <- value
}
\arguments{
  \item{rbm}{A instance of the class \code{\link{RBM}}.}

  \item{value}{The weights of the \code{\link{RBM}}
  object.}
}
\description{
Sets the weights of the \code{\link{RBM}} object
}
\seealso{
\code{\link{RBM}}
}

