% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{sigmoidUnitDerivative}
\alias{sigmoidUnitDerivative}
\title{Sigmoid unit function with unit derivatives.}
\usage{
sigmoidUnitDerivative(input, ...)
}
\arguments{
\item{input}{Input for the activation function.}
}
\value{
A list with the activation in the first entry and the derivative of
  the transfer function in the second entry
}
\description{
The function calculates the activation and returns a list which the first
entry is the result through the sigmoid transfer function and the second
entry is the derivative of the transfer function.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{linearUnitDerivative}};
  \code{\link{maxoutUnitDerivative}};
  \code{\link{rectifiedLinearUnitDerivative}};
  \code{\link{softmaxUnitDerivative}};
  \code{\link{tanSigmoidUnitDerivative}}
}

