% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darchUnitFunctions.R
\name{tanSigmoidUnitDerivative}
\alias{tanSigmoidUnitDerivative}
\title{Continuous Tan-Sigmoid unit function.}
\usage{
tanSigmoidUnitDerivative(input, ...)
}
\arguments{
\item{input}{Input for the activation function.}
}
\value{
A list with the activation in the first entry and the derivative of
  the transfer function in the second entry
}
\description{
Calculates the unit activations and returns them in a list.
}
\seealso{
\code{\linkS4class{DArch}}

Other DArch unit functions: \code{\link{linearUnitDerivative}};
  \code{\link{maxoutUnitDerivative}};
  \code{\link{rectifiedLinearUnitDerivative}};
  \code{\link{sigmoidUnitDerivative}};
  \code{\link{softmaxUnitDerivative}}
}

